Extract Data from Daymet.

Extracts all available variables from Daymet and matches them to input observation data. All variables are available at a daily resolution since 1980 in North America and Hawaii, and since 1950 in Puerto Rico, and at a ~ 1 km spatial resolution. This data can be requested from the NASA AppEEARS service using daymet_request() and downloaded using daymet_download().

Usage

daymet_extract(
  data,
  daymet_reqs,
  covariates = "daymet_prcp",
  site_name = NULL,
  date_year = NULL,
  date_month = NULL,
  date_day = NULL,
  dl_path = NULL,
  verbose = TRUE,
  retain = TRUE
)

Arguments

data

A data.frame, sf 'POINT' or 'POLYGON' object, or terra 'points' or 'polygons' object containing columns with the year, month, and day an observation was made either named the BMDE defaults survey_year, survey_month , and survey_day respectively or another name specified in arguments date_year, date_month, and/or date_day.

daymet_reqs

data.frame. A data.frame with columns 1) request_name containing AppEEARS request names, 2) request_id containing AppEEARS request IDs, and optionally 3) date containing the date for which the associated request is downloading data for, or a filepath to a .rds file containing such data. The direct output of daymet_request() can be supplied here.

covariates

Character, vector if multiple Daymet data types desired. By default, extracts Daymet precipitation data.

site_name

Character. Optional argument to provide the name of the column containing site names if not contained within the BMDE column SurveyAreaIdentifier. Can be left NULL and still function properly if originally specified in a call to data_fmt() or daymet_download().

date_year

Character. Optional argument to provide the name of the column containing year data if not contained within the BMDE column survey_year. Can be left NULL and still function properly if originally specified in a call to data_fmt() or daymet_download().

date_month

Character. Optional argument to provide the name of the column containing month data if not contained within the BMDE column survey_month. Can be left NULL and still function properly if originally specified in a call to data_fmt() or daymet_download().

date_day

Character. Optional argument to provide the name of the column containing day-of-month (i.e., a number from 1 to 31) data if not contained within the BMDE column survey_day. Can be left NULL and still function properly if originally specified in a call to data_fmt() or daymet_download().

dl_path

Character. Optional argument to provide path to downloaded data. By default, data is downloaded to a subfolder daymet/ in the working directory.

verbose

Logical. Should messages be displayed?

retain

Logical. Should Daymet data files be kept after extraction? If FALSE, files will be deleted.

Value

For sf 'POINT' or terra 'points' input data, original data with column(s) appended containing the Daymet data value(s) at each point.

For sf 'POLYGON' or terra 'polygons' input data, original data with column(s) appended containing the mean Daymet data value(s) within each polygon.

Details

One (or multiple) Daymet variable(s) can be extracted by specifying the following values to the covariates argument. The appropriate variables must be available in the AppEEARS request supplied to daymet_reqs. Requests can be submitted via daymet_request() and downloaded via daymet_download():

Day length (s/day): daymet_dayl

Precipitation (mm/day): daymet_prcp

Shortwave radiation (W/m^2): dayment_srad

Snow water equivalent (kg/m^2): daymet_swe

Maximum air temperature (°C): daymet_tmax

Minimum air temperature (°C): daymet_tmin

Water vapor pressure (Pa): daymet_vp

See also

Examples