Load Terrain Tiles Elevation Data

Loads Mapzen Terrain Tiles elevation data, delivered at varying spatial resolutions. This data is open access via Amazon Web Services, and is a global composite of a variety of data sources. Data is loaded into the R environment, and is not permanently downloaded onto the user's operating system.

Usage

elevation_download(
  data,
  site_name = NULL,
  negative_to_na = FALSE,
  z = 7,
  dl_path = NULL,
  progress = TRUE
)

Arguments

data

An sf 'POINT' or 'POLYGON' object, or terra 'points' or 'polygons' object.

site_name

Character. Optional argument to provide the name of the column containing site names if not contained within the BMDE column SurveyAreaIdentifier. Can be left NULL and still function properly if originally specified in a call to data_fmt().

negative_to_na

Logical. Should negative elevation values be turned into NA? If FALSE, check the outputted rasters for unrealistically large negative values as these may instead represent missing data. For coastal areas, negative elevation values may represent bathymetry (i.e. sea depth).

z

Numeric. Zoom level to fetch, determining the resulting spatial resolution of downloaded elevation data. More information can be found here.

dl_path

Character. Optional argument to provide path to download data to. By default, data is downloaded to a subfolder TerrainTiles/ in the working directory.

progress

Logical. Should progress bars and download messages be displayed?

Value

A terra SpatRaster in the projection of the data supplied to the data argument, covering the bounding box of the supplied data.

Details

Users should be conscious of the final spatial resolution of their elevation data, as this varies by latitude and zoom level. This can be accessed using terra::res().

Downloads are facilitated by a call to elevatr::get_elev_raster().

See also

Examples

# Convert included, test data on black-capped chickadees to sf POINT object
bcch <- sf::st_as_sf(
  bcch,
  coords = c("longitude", "latitude"),
  crs = 4326
)

# Load Terrain Tiles data
output <- elevation_download(data = bcch)
#> [Elevation Download] downloading data.
#> Mosaicing & Projecting
#> Note: Elevation units are in meters.