Metadata

These functions return metadata codes, names, descriptions, and information associated with the data downloaded from NatureCounts.

meta_country_codes()

meta_statprov_codes()

meta_subnational2_codes()

meta_iba_codes()

meta_bcr_codes()

meta_utm_squares()

meta_species_authority()

meta_species_codes()

meta_species_taxonomy()

meta_collections()

meta_breeding_codes()

meta_project_protocols()

meta_projects()

meta_protocol_types()

meta_bmde_versions()

meta_bmde_fields(version = "minimum")

Arguments

version
Character. BMDE version for which to return fields

Value

Data frame

Details

Some of these metadata are stored locally and can be updated with the nc_metadata() function. Others are downloaded as requested.

Functions

meta_country_codes: Country codes

meta_statprov_codes: State/Province codes

meta_subnational2_codes: Subnational2 codes

meta_iba_codes: Important Bird Area (IBA) codes

meta_bcr_codes: Bird Conservation Region (BCR) codes

meta_utm_squares: UTM Square codes

meta_species_authority: Species taxonomic authorities

meta_species_codes: Alpha-numeric codes for avian species

meta_species_taxonomy: Codes and taxonomic information for all species

meta_collections: Collections names and descriptions

meta_breeding_codes: Breeding codes and descriptions

meta_project_protocols: Project protocols

meta_projects: Projects ids, names, websites, and descriptions

meta_protocol_types: Protocol types and descriptions

meta_bmde_versions: Names and descriptions of the available versions of BMDE (Bird Monitoring Data Exchange). These refer to sets of fields/columns which can be downloaded for a given group of data. See nc_data_dl() for more details.

meta_bmde_fields: Fields/columns associated with a particular BMDE (Bird Monitoring Data Exchange) version. See meta_bmde_versions() for the different versions available, meta_collections() for which version is used by which project, and nc_data_dl() for more details on downloading data with a given set of fields/columns.

Examples

# Return fields/columns in the 'minimum' version
meta_bmde_fields()

#>    field_order      version              local_name              field_name
#> 1            1 BMDE2.00-min  GlobalUniqueIdentifier  GlobalUniqueIdentifier
#> 2            2 BMDE2.00-min           CatalogNumber           CatalogNumber
#> 3            3 BMDE2.00-min                Locality                Locality
#> 4            4 BMDE2.00-min           TimeCollected           TimeCollected
#> 5            5 BMDE2.00-min         CollectorNumber         CollectorNumber
#> 6            6 BMDE2.00-min             FieldNumber             FieldNumber
#> 7            7 BMDE2.00-min                 Remarks                 Remarks
#> 8            8 BMDE2.00-min             ProjectCode             ProjectCode
#> 9            9 BMDE2.00-min            ProtocolType            ProtocolType
#> 10          10 BMDE2.00-min            ProtocolCode            ProtocolCode
#> 11          11 BMDE2.00-min             ProtocolURL             ProtocolURL
#> 12          12 BMDE2.00-min    SurveyAreaIdentifier    SurveyAreaIdentifier
#> 13          13 BMDE2.00-min SamplingEventIdentifier SamplingEventIdentifier
#> 14          14 BMDE2.00-min  SamplingEventStructure  SamplingEventStructure
#> 15          15 BMDE2.00-min         RouteIdentifier         RouteIdentifier
#> 16          16 BMDE2.00-min TimeObservationsStarted TimeObservationsStarted
#> 17          17 BMDE2.00-min   TimeObservationsEnded   TimeObservationsEnded
#> 18          18 BMDE2.00-min         DurationInHours         DurationInHours
#> 19          19 BMDE2.00-min     TimeIntervalStarted     TimeIntervalStarted
#> 20          20 BMDE2.00-min       TimeIntervalEnded       TimeIntervalEnded
#> 21          21 BMDE2.00-min   TimeIntervalsAdditive   TimeIntervalsAdditive
#> 22          22 BMDE2.00-min       NumberOfObservers       NumberOfObservers
#> 23          23 BMDE2.00-min          NoObservations          NoObservations
#> 24          24 BMDE2.00-min        ObservationCount        ObservationCount
#> 25          25 BMDE2.00-min   ObservationDescriptor   ObservationDescriptor
#> 26          26 BMDE2.00-min       ObservationCount2       ObservationCount2
#> 27          27 BMDE2.00-min  ObservationDescriptor2  ObservationDescriptor2
#> 28          28 BMDE2.00-min       ObservationCount3       ObservationCount3
#> 29          29 BMDE2.00-min  ObservationDescriptor3  ObservationDescriptor3
#> 30          30 BMDE2.00-min       ObservationCount4       ObservationCount4
#> 31          31 BMDE2.00-min  ObservationDescriptor4  ObservationDescriptor4
#> 32          32 BMDE2.00-min       ObservationCount5       ObservationCount5
#> 33          33 BMDE2.00-min  ObservationDescriptor5  ObservationDescriptor5
#> 34          34 BMDE2.00-min       ObservationCount6       ObservationCount6
#> 35          35 BMDE2.00-min  ObservationDescriptor6  ObservationDescriptor6
#> 36          36 BMDE2.00-min  AllIndividualsReported  AllIndividualsReported
#> 37          37 BMDE2.00-min      AllSpeciesReported      AllSpeciesReported

# Retrun fields/columns in the 'core' version
meta_bmde_fields(version = "core")

#>     field_order  version                             local_name
#> 1             1 BMDE2.00                 GlobalUniqueIdentifier
#> 2             2 BMDE2.00                       DateLastModified
#> 3             3 BMDE2.00                          BasisOfRecord
#> 4             4 BMDE2.00                        InstitutionCode
#> 5             5 BMDE2.00                         CollectionCode
#> 6             6 BMDE2.00                          CatalogNumber
#> 7             7 BMDE2.00                         ScientificName
#> 8             8 BMDE2.00                            HigherTaxon
#> 9             9 BMDE2.00                                Kingdom
#> 10           10 BMDE2.00                                 Phylum
#> 11           11 BMDE2.00                                  Class
#> 12           12 BMDE2.00                             OrderTaxon
#> 13           13 BMDE2.00                                 Family
#> 14           14 BMDE2.00                                  Genus
#> 15           15 BMDE2.00                        SpecificEpithet
#> 16           16 BMDE2.00                      InfraspecificRank
#> 17           17 BMDE2.00                   InfraspecificEpithet
#> 18           18 BMDE2.00                   ScientificNameAuthor
#> 19           19 BMDE2.00                IdentificationQualifier
#> 20           20 BMDE2.00                        HigherGeography
#> 21           21 BMDE2.00                              Continent
#> 22           22 BMDE2.00                              WaterBody
#> 23           23 BMDE2.00                            IslandGroup
#> 24           24 BMDE2.00                                 Island
#> 25           25 BMDE2.00                                Country
#> 26           26 BMDE2.00                          StateProvince
#> 27           27 BMDE2.00                                 County
#> 28           28 BMDE2.00                               Locality
#> 29           29 BMDE2.00               MinimumElevationInMeters
#> 30           30 BMDE2.00               MaximumElevationInMeters
#> 31           31 BMDE2.00                   MinimumDepthInMeters
#> 32           32 BMDE2.00                   MaximumDepthInMeters
#> 33           33 BMDE2.00                        DecimalLatitude
#> 34           34 BMDE2.00                       DecimalLongitude
#> 35           35 BMDE2.00                          GeodeticDatum
#> 36           36 BMDE2.00          CoordinateUncertaintyInMeters
#> 37           37 BMDE2.00                          YearCollected
#> 38           38 BMDE2.00                         MonthCollected
#> 39           39 BMDE2.00                           DayCollected
#> 40           40 BMDE2.00                          TimeCollected
#> 41           41 BMDE2.00                              JulianDay
#> 42           42 BMDE2.00                              Collector
#> 43           43 BMDE2.00                                    Sex
#> 44           44 BMDE2.00                              LifeStage
#> 45           45 BMDE2.00                               ImageURL
#> 46           46 BMDE2.00                     RelatedInformation
#> 47           47 BMDE2.00                        CollectorNumber
#> 48           48 BMDE2.00                            FieldNumber
#> 49           49 BMDE2.00                             FieldNotes
#> 50           50 BMDE2.00              OriginalCoordinatesSystem
#> 51           51 BMDE2.00                        LatLongComments
#> 52           52 BMDE2.00                     GeoreferenceMethod
#> 53           53 BMDE2.00                 GeoreferenceReferences
#> 54           54 BMDE2.00         GeoreferenceVerificationStatus
#> 55           55 BMDE2.00                                Remarks
#> 56           56 BMDE2.00                           FootprintWKT
#> 57           57 BMDE2.00                           FootprintSRS
#> 58           58 BMDE2.00                            ProjectCode
#> 59           59 BMDE2.00                           ProtocolType
#> 60           60 BMDE2.00                           ProtocolCode
#> 61           61 BMDE2.00                ProtocolSpeciesTargeted
#> 62           62 BMDE2.00                      ProtocolReference
#> 63           63 BMDE2.00                            ProtocolURL
#> 64           64 BMDE2.00                   SurveyAreaIdentifier
#> 65           65 BMDE2.00                         SurveyAreaSize
#> 66           66 BMDE2.00            SurveyAreaPercentageCovered
#> 67           67 BMDE2.00                        SurveyAreaShape
#> 68           68 BMDE2.00               SurveyAreaLongAxisLength
#> 69           69 BMDE2.00              SurveyAreaShortAxisLength
#> 70           70 BMDE2.00          SurveyAreaLongAxisOrientation
#> 71           71 BMDE2.00                       CoordinatesScope
#> 72           72 BMDE2.00                SamplingEventIdentifier
#> 73           73 BMDE2.00                 SamplingEventStructure
#> 74           74 BMDE2.00                        RouteIdentifier
#> 75           75 BMDE2.00                TimeObservationsStarted
#> 76           76 BMDE2.00                  TimeObservationsEnded
#> 77           77 BMDE2.00                        DurationInHours
#> 78           78 BMDE2.00                    TimeIntervalStarted
#> 79           79 BMDE2.00                      TimeIntervalEnded
#> 80           80 BMDE2.00                  TimeIntervalsAdditive
#> 81           81 BMDE2.00                      NumberOfObservers
#> 82           82 BMDE2.00                     EffortMeasurement1
#> 83           83 BMDE2.00                           EffortUnits1
#> 84           84 BMDE2.00                     EffortMeasurement2
#> 85           85 BMDE2.00                           EffortUnits2
#> 86           86 BMDE2.00                     EffortMeasurement3
#> 87           87 BMDE2.00                           EffortUnits3
#> 88           88 BMDE2.00                     EffortMeasurement4
#> 89           89 BMDE2.00                           EffortUnits4
#> 90           90 BMDE2.00                     EffortMeasurement5
#> 91           91 BMDE2.00                           EffortUnits5
#> 92           92 BMDE2.00                     EffortMeasurement6
#> 93           93 BMDE2.00                           EffortUnits6
#> 94           94 BMDE2.00                     EffortMeasurement7
#> 95           95 BMDE2.00                           EffortUnits7
#> 96           96 BMDE2.00                     EffortMeasurement8
#> 97           97 BMDE2.00                           EffortUnits8
#> 98           98 BMDE2.00                     EffortMeasurement9
#> 99           99 BMDE2.00                           EffortUnits9
#> 100         100 BMDE2.00                    EffortMeasurement10
#> 101         101 BMDE2.00                          EffortUnits10
#> 102         102 BMDE2.00                    EffortMeasurement11
#> 103         103 BMDE2.00                          EffortUnits11
#> 104         104 BMDE2.00                    EffortMeasurement12
#> 105         105 BMDE2.00                          EffortUnits12
#> 106         106 BMDE2.00                    EffortMeasurement13
#> 107         107 BMDE2.00                          EffortUnits13
#> 108         108 BMDE2.00                    EffortMeasurement14
#> 109         109 BMDE2.00                          EffortUnits14
#> 110         110 BMDE2.00                    EffortMeasurement15
#> 111         111 BMDE2.00                          EffortUnits15
#> 112         112 BMDE2.00                    EffortMeasurement16
#> 113         113 BMDE2.00                          EffortUnits16
#> 114         114 BMDE2.00                    EffortMeasurement17
#> 115         115 BMDE2.00                          EffortUnits17
#> 116         116 BMDE2.00                    EffortMeasurement18
#> 117         117 BMDE2.00                          EffortUnits18
#> 118         118 BMDE2.00                         NoObservations
#> 119         119 BMDE2.00                   DistanceFromObserver
#> 120         120 BMDE2.00                DistanceFromObserverMin
#> 121         121 BMDE2.00                DistanceFromObserverMax
#> 122         122 BMDE2.00                      DistanceFromStart
#> 123         123 BMDE2.00                       BearingInDegrees
#> 124         124 BMDE2.00                SpecimenDecimalLatitude
#> 125         125 BMDE2.00               SpecimenDecimalLongitude
#> 126         126 BMDE2.00                  SpecimenGeodeticDatum
#> 127         127 BMDE2.00                        SpecimenUTMZone
#> 128         128 BMDE2.00                    SpecimenUTMNorthing
#> 129         129 BMDE2.00                     SpecimenUTMEasting
#> 130         130 BMDE2.00                       ObservationCount
#> 131         131 BMDE2.00                  ObservationDescriptor
#> 132         132 BMDE2.00                      ObservationCount2
#> 133         133 BMDE2.00                 ObservationDescriptor2
#> 134         134 BMDE2.00                      ObservationCount3
#> 135         135 BMDE2.00                 ObservationDescriptor3
#> 136         136 BMDE2.00                      ObservationCount4
#> 137         137 BMDE2.00                 ObservationDescriptor4
#> 138         138 BMDE2.00                      ObservationCount5
#> 139         139 BMDE2.00                 ObservationDescriptor5
#> 140         140 BMDE2.00                      ObservationCount6
#> 141         141 BMDE2.00                 ObservationDescriptor6
#> 142         142 BMDE2.00                        ObsCountAtLeast
#> 143         143 BMDE2.00                         ObsCountAtMost
#> 144         144 BMDE2.00                        ObservationDate
#> 145         145 BMDE2.00                  DateUncertaintyInDays
#> 146         146 BMDE2.00                 AllIndividualsReported
#> 147         147 BMDE2.00                     AllSpeciesReported
#> 148         148 BMDE2.00                                UTMZone
#> 149         149 BMDE2.00                            UTMNorthing
#> 150         150 BMDE2.00                             UTMEasting
#> 151         151 BMDE2.00 CoordinatesUncertaintyInDecimalDegrees
#> 152         152 BMDE2.00                             CommonName
#> 153         153 BMDE2.00                      RecordPermissions
#> 154         154 BMDE2.00                   MultiScientificName1
#> 155         155 BMDE2.00                   MultiScientificName2
#> 156         156 BMDE2.00                   MultiScientificName3
#> 157         157 BMDE2.00                   MultiScientificName4
#> 158         158 BMDE2.00                   MultiScientificName5
#> 159         159 BMDE2.00                   MultiScientificName6
#> 160         160 BMDE2.00              TaxonomicAuthorityAuthors
#> 161         161 BMDE2.00              TaxonomicAuthorityVersion
#> 162         162 BMDE2.00                 TaxonomicAuthorityYear
#> 163         163 BMDE2.00                            SpeciesCode
#> 164         164 BMDE2.00                         TaxonConceptID
#> 165         165 BMDE2.00                  BreedingBirdAtlasCode
#> 166         166 BMDE2.00                     HabitatDescription
#> 167         167 BMDE2.00                               Remarks2
#> 168         168 BMDE2.00                     LastModifiedAction
#> 169         169 BMDE2.00                     RecordReviewStatus
#>                                 field_name
#> 1                   GlobalUniqueIdentifier
#> 2                         DateLastModified
#> 3                            BasisOfRecord
#> 4                          InstitutionCode
#> 5                           CollectionCode
#> 6                            CatalogNumber
#> 7                           ScientificName
#> 8                              HigherTaxon
#> 9                                  Kingdom
#> 10                                  Phylum
#> 11                                   Class
#> 12                                   Order
#> 13                                  Family
#> 14                                   Genus
#> 15                         SpecificEpithet
#> 16                       InfraspecificRank
#> 17                    InfraspecificEpithet
#> 18                    ScientificNameAuthor
#> 19                 IdentificationQualifier
#> 20                         HigherGeography
#> 21                               Continent
#> 22                               WaterBody
#> 23                             IslandGroup
#> 24                                  Island
#> 25                                 Country
#> 26                           StateProvince
#> 27                                  County
#> 28                                Locality
#> 29                MinimumElevationInMeters
#> 30                MaximumElevationInMeters
#> 31                    MinimumDepthInMeters
#> 32                    MaximumDepthInMeters
#> 33                         DecimalLatitude
#> 34                        DecimalLongitude
#> 35                           GeodeticDatum
#> 36           CoordinateUncertaintyInMeters
#> 37                           YearCollected
#> 38                          MonthCollected
#> 39                            DayCollected
#> 40                           TimeCollected
#> 41                               JulianDay
#> 42                               Collector
#> 43                                     Sex
#> 44                               LifeStage
#> 45                                ImageURL
#> 46                      RelatedInformation
#> 47                         CollectorNumber
#> 48                             FieldNumber
#> 49                              FieldNotes
#> 50               OriginalCoordinatesSystem
#> 51                         LatLongComments
#> 52                      GeoreferenceMethod
#> 53                  GeoreferenceReferences
#> 54          GeoreferenceVerificationStatus
#> 55                                 Remarks
#> 56                            FootprintWKT
#> 57                            FootprintSRS
#> 58                             ProjectCode
#> 59                            ProtocolType
#> 60                            ProtocolCode
#> 61                 ProtocolSpeciesTargeted
#> 62                       ProtocolReference
#> 63                             ProtocolURL
#> 64                    SurveyAreaIdentifier
#> 65                          SurveyAreaSize
#> 66             SurveyAreaPercentageCovered
#> 67                         SurveyAreaShape
#> 68                SurveyAreaLongAxisLength
#> 69               SurveyAreaShortAxisLength
#> 70           SurveyAreaLongAxisOrientation
#> 71                        CoordinatesScope
#> 72                 SamplingEventIdentifier
#> 73                  SamplingEventStructure
#> 74                         RouteIdentifier
#> 75                 TimeObservationsStarted
#> 76                   TimeObservationsEnded
#> 77                         DurationInHours
#> 78                     TimeIntervalStarted
#> 79                       TimeIntervalEnded
#> 80                   TimeIntervalsAdditive
#> 81                       NumberOfObservers
#> 82                      EffortMeasurement1
#> 83                            EffortUnits1
#> 84                      EffortMeasurement2
#> 85                            EffortUnits2
#> 86                      EffortMeasurement3
#> 87                            EffortUnits3
#> 88                      EffortMeasurement4
#> 89                            EffortUnits4
#> 90                      EffortMeasurement5
#> 91                            EffortUnits5
#> 92                      EffortMeasurement6
#> 93                            EffortUnits6
#> 94                      EffortMeasurement7
#> 95                            EffortUnits7
#> 96                      EffortMeasurement8
#> 97                            EffortUnits8
#> 98                      EffortMeasurement9
#> 99                            EffortUnits9
#> 100                    EffortMeasurement10
#> 101                          EffortUnits10
#> 102                    EffortMeasurement11
#> 103                          EffortUnits11
#> 104                    EffortMeasurement12
#> 105                          EffortUnits12
#> 106                    EffortMeasurement13
#> 107                          EffortUnits13
#> 108                    EffortMeasurement14
#> 109                          EffortUnits14
#> 110                    EffortMeasurement15
#> 111                          EffortUnits15
#> 112                    EffortMeasurement16
#> 113                          EffortUnits16
#> 114                    EffortMeasurement17
#> 115                          EffortUnits17
#> 116                    EffortMeasurement18
#> 117                          EffortUnits18
#> 118                         NoObservations
#> 119                   DistanceFromObserver
#> 120                DistanceFromObserverMin
#> 121                DistanceFromObserverMax
#> 122                      DistanceFromStart
#> 123                       BearingInDegrees
#> 124                SpecimenDecimalLatitude
#> 125               SpecimenDecimalLongitude
#> 126                  SpecimenGeodeticDatum
#> 127                        SpecimenUTMZone
#> 128                    SpecimenUTMNorthing
#> 129                     SpecimenUTMEasting
#> 130                       ObservationCount
#> 131                  ObservationDescriptor
#> 132                      ObservationCount2
#> 133                 ObservationDescriptor2
#> 134                      ObservationCount3
#> 135                 ObservationDescriptor3
#> 136                      ObservationCount4
#> 137                 ObservationDescriptor4
#> 138                      ObservationCount5
#> 139                 ObservationDescriptor5
#> 140                      ObservationCount6
#> 141                 ObservationDescriptor6
#> 142                        ObsCountAtLeast
#> 143                         ObsCountAtMost
#> 144                        ObservationDate
#> 145                  DateUncertaintyInDays
#> 146                 AllIndividualsReported
#> 147                     AllSpeciesReported
#> 148                                UTMZone
#> 149                            UTMNorthing
#> 150                             UTMEasting
#> 151 CoordinatesUncertaintyInDecimalDegrees
#> 152                             CommonName
#> 153                      RecordPermissions
#> 154                   MultiScientificName1
#> 155                   MultiScientificName2
#> 156                   MultiScientificName3
#> 157                   MultiScientificName4
#> 158                   MultiScientificName5
#> 159                   MultiScientificName6
#> 160              TaxonomicAuthorityAuthors
#> 161              TaxonomicAuthorityVersion
#> 162                 TaxonomicAuthorityYear
#> 163                            SpeciesCode
#> 164                         TaxonConceptID
#> 165                  BreedingBirdAtlasCode
#> 166                     HabitatDescription
#> 167                               Remarks2
#> 168                     LastModifiedAction
#> 169                     RecordReviewStatus

# Return all possible fields
meta_bmde_fields(version = "extended")

#>     field_order      version                             local_name
#> 1             1 BMDE2.00-ext                 GlobalUniqueIdentifier
#> 2             2 BMDE2.00-ext                       DateLastModified
#> 3             3 BMDE2.00-ext                          BasisOfRecord
#> 4             4 BMDE2.00-ext                          InstitutionID
#> 5             5 BMDE2.00-ext                        InstitutionCode
#> 6             6 BMDE2.00-ext                         CollectionCode
#> 7             7 BMDE2.00-ext                          CatalogNumber
#> 8             8 BMDE2.00-ext                         ScientificName
#> 9             9 BMDE2.00-ext                            HigherTaxon
#> 10           10 BMDE2.00-ext                                Kingdom
#> 11           11 BMDE2.00-ext                                 Phylum
#> 12           12 BMDE2.00-ext                                  Class
#> 13           13 BMDE2.00-ext                             OrderTaxon
#> 14           14 BMDE2.00-ext                                 Family
#> 15           15 BMDE2.00-ext                                  Genus
#> 16           16 BMDE2.00-ext                        SpecificEpithet
#> 17           17 BMDE2.00-ext                      InfraspecificRank
#> 18           18 BMDE2.00-ext                   InfraspecificEpithet
#> 19           19 BMDE2.00-ext                   ScientificNameAuthor
#> 20           20 BMDE2.00-ext                IdentificationQualifier
#> 21           21 BMDE2.00-ext                        HigherGeography
#> 22           22 BMDE2.00-ext                              Continent
#> 23           23 BMDE2.00-ext                              WaterBody
#> 24           24 BMDE2.00-ext                            IslandGroup
#> 25           25 BMDE2.00-ext                                 Island
#> 26           26 BMDE2.00-ext                                Country
#> 27           27 BMDE2.00-ext                          StateProvince
#> 28           28 BMDE2.00-ext                                 County
#> 29           29 BMDE2.00-ext                               Locality
#> 30           30 BMDE2.00-ext               MinimumElevationInMeters
#> 31           31 BMDE2.00-ext               MaximumElevationInMeters
#> 32           32 BMDE2.00-ext                   MinimumDepthInMeters
#> 33           33 BMDE2.00-ext                   MaximumDepthInMeters
#> 34           34 BMDE2.00-ext                        DecimalLatitude
#> 35           35 BMDE2.00-ext                       DecimalLongitude
#> 36           36 BMDE2.00-ext                          GeodeticDatum
#> 37           37 BMDE2.00-ext          CoordinateUncertaintyInMeters
#> 38           38 BMDE2.00-ext                          YearCollected
#> 39           39 BMDE2.00-ext                         MonthCollected
#> 40           40 BMDE2.00-ext                           DayCollected
#> 41           41 BMDE2.00-ext                          TimeCollected
#> 42           42 BMDE2.00-ext                              JulianDay
#> 43           43 BMDE2.00-ext                              Collector
#> 44           44 BMDE2.00-ext                                    Sex
#> 45           45 BMDE2.00-ext                              LifeStage
#> 46           46 BMDE2.00-ext                               ImageURL
#> 47           47 BMDE2.00-ext                     RelatedInformation
#> 48           48 BMDE2.00-ext                        CollectorNumber
#> 49           49 BMDE2.00-ext                            FieldNumber
#> 50           50 BMDE2.00-ext                             FieldNotes
#> 51           51 BMDE2.00-ext              OriginalCoordinatesSystem
#> 52           52 BMDE2.00-ext                        LatLongComments
#> 53           53 BMDE2.00-ext                     GeoreferenceMethod
#> 54           54 BMDE2.00-ext                 GeoreferenceReferences
#> 55           55 BMDE2.00-ext         GeoreferenceVerificationStatus
#> 56           56 BMDE2.00-ext                                Remarks
#> 57           57 BMDE2.00-ext                           FootprintWKT
#> 58           58 BMDE2.00-ext                           FootprintSRS
#> 59           59 BMDE2.00-ext                            ProjectCode
#> 60           60 BMDE2.00-ext                           ProtocolType
#> 61           61 BMDE2.00-ext                           ProtocolCode
#> 62           62 BMDE2.00-ext                ProtocolSpeciesTargeted
#> 63           63 BMDE2.00-ext                      ProtocolReference
#> 64           64 BMDE2.00-ext                            ProtocolURL
#> 65           65 BMDE2.00-ext                   SurveyAreaIdentifier
#> 66           66 BMDE2.00-ext                         SurveyAreaSize
#> 67           67 BMDE2.00-ext            SurveyAreaPercentageCovered
#> 68           68 BMDE2.00-ext                        SurveyAreaShape
#> 69           69 BMDE2.00-ext               SurveyAreaLongAxisLength
#> 70           70 BMDE2.00-ext              SurveyAreaShortAxisLength
#> 71           71 BMDE2.00-ext          SurveyAreaLongAxisOrientation
#> 72           72 BMDE2.00-ext                       CoordinatesScope
#> 73           73 BMDE2.00-ext                SamplingEventIdentifier
#> 74           74 BMDE2.00-ext                 SamplingEventStructure
#> 75           75 BMDE2.00-ext                        RouteIdentifier
#> 76           76 BMDE2.00-ext                TimeObservationsStarted
#> 77           77 BMDE2.00-ext                  TimeObservationsEnded
#> 78           78 BMDE2.00-ext                        DurationInHours
#> 79           79 BMDE2.00-ext                    TimeIntervalStarted
#> 80           80 BMDE2.00-ext                      TimeIntervalEnded
#> 81           81 BMDE2.00-ext                  TimeIntervalsAdditive
#> 82           82 BMDE2.00-ext                      NumberOfObservers
#> 83           83 BMDE2.00-ext                     EffortMeasurement1
#> 84           84 BMDE2.00-ext                           EffortUnits1
#> 85           85 BMDE2.00-ext                     EffortMeasurement2
#> 86           86 BMDE2.00-ext                           EffortUnits2
#> 87           87 BMDE2.00-ext                     EffortMeasurement3
#> 88           88 BMDE2.00-ext                           EffortUnits3
#> 89           89 BMDE2.00-ext                     EffortMeasurement4
#> 90           90 BMDE2.00-ext                           EffortUnits4
#> 91           91 BMDE2.00-ext                     EffortMeasurement5
#> 92           92 BMDE2.00-ext                           EffortUnits5
#> 93           93 BMDE2.00-ext                     EffortMeasurement6
#> 94           94 BMDE2.00-ext                           EffortUnits6
#> 95           95 BMDE2.00-ext                     EffortMeasurement7
#> 96           96 BMDE2.00-ext                           EffortUnits7
#> 97           97 BMDE2.00-ext                     EffortMeasurement8
#> 98           98 BMDE2.00-ext                           EffortUnits8
#> 99           99 BMDE2.00-ext                     EffortMeasurement9
#> 100         100 BMDE2.00-ext                           EffortUnits9
#> 101         101 BMDE2.00-ext                    EffortMeasurement10
#> 102         102 BMDE2.00-ext                          EffortUnits10
#> 103         103 BMDE2.00-ext                    EffortMeasurement11
#> 104         104 BMDE2.00-ext                          EffortUnits11
#> 105         105 BMDE2.00-ext                    EffortMeasurement12
#> 106         106 BMDE2.00-ext                          EffortUnits12
#> 107         107 BMDE2.00-ext                    EffortMeasurement13
#> 108         108 BMDE2.00-ext                          EffortUnits13
#> 109         109 BMDE2.00-ext                    EffortMeasurement14
#> 110         110 BMDE2.00-ext                          EffortUnits14
#> 111         111 BMDE2.00-ext                    EffortMeasurement15
#> 112         112 BMDE2.00-ext                          EffortUnits15
#> 113         113 BMDE2.00-ext                    EffortMeasurement16
#> 114         114 BMDE2.00-ext                          EffortUnits16
#> 115         115 BMDE2.00-ext                    EffortMeasurement17
#> 116         116 BMDE2.00-ext                          EffortUnits17
#> 117         117 BMDE2.00-ext                    EffortMeasurement18
#> 118         118 BMDE2.00-ext                          EffortUnits18
#> 119         119 BMDE2.00-ext                         NoObservations
#> 120         120 BMDE2.00-ext                   DistanceFromObserver
#> 121         121 BMDE2.00-ext                DistanceFromObserverMin
#> 122         122 BMDE2.00-ext                DistanceFromObserverMax
#> 123         123 BMDE2.00-ext                      DistanceFromStart
#> 124         124 BMDE2.00-ext                       BearingInDegrees
#> 125         125 BMDE2.00-ext                SpecimenDecimalLatitude
#> 126         126 BMDE2.00-ext               SpecimenDecimalLongitude
#> 127         127 BMDE2.00-ext                  SpecimenGeodeticDatum
#> 128         128 BMDE2.00-ext                        SpecimenUTMZone
#> 129         129 BMDE2.00-ext                    SpecimenUTMNorthing
#> 130         130 BMDE2.00-ext                     SpecimenUTMEasting
#> 131         131 BMDE2.00-ext                       ObservationCount
#> 132         132 BMDE2.00-ext                  ObservationDescriptor
#> 133         133 BMDE2.00-ext                      ObservationCount2
#> 134         134 BMDE2.00-ext                 ObservationDescriptor2
#> 135         135 BMDE2.00-ext                      ObservationCount3
#> 136         136 BMDE2.00-ext                 ObservationDescriptor3
#> 137         137 BMDE2.00-ext                      ObservationCount4
#> 138         138 BMDE2.00-ext                 ObservationDescriptor4
#> 139         139 BMDE2.00-ext                      ObservationCount5
#> 140         140 BMDE2.00-ext                 ObservationDescriptor5
#> 141         141 BMDE2.00-ext                      ObservationCount6
#> 142         142 BMDE2.00-ext                 ObservationDescriptor6
#> 143         143 BMDE2.00-ext                      ObservationCount7
#> 144         144 BMDE2.00-ext                 ObservationDescriptor7
#> 145         145 BMDE2.00-ext                      ObservationCount8
#> 146         146 BMDE2.00-ext                 ObservationDescriptor8
#> 147         147 BMDE2.00-ext                      ObservationCount9
#> 148         148 BMDE2.00-ext                 ObservationDescriptor9
#> 149         149 BMDE2.00-ext                     ObservationCount10
#> 150         150 BMDE2.00-ext                ObservationDescriptor10
#> 151         151 BMDE2.00-ext                     ObservationCount11
#> 152         152 BMDE2.00-ext                ObservationDescriptor11
#> 153         153 BMDE2.00-ext                     ObservationCount12
#> 154         154 BMDE2.00-ext                ObservationDescriptor12
#> 155         155 BMDE2.00-ext                     ObservationCount13
#> 156         156 BMDE2.00-ext                ObservationDescriptor13
#> 157         157 BMDE2.00-ext                     ObservationCount14
#> 158         158 BMDE2.00-ext                ObservationDescriptor14
#> 159         159 BMDE2.00-ext                        ObsCountAtLeast
#> 160         160 BMDE2.00-ext                         ObsCountAtMost
#> 161         161 BMDE2.00-ext                        ObservationDate
#> 162         162 BMDE2.00-ext                  DateUncertaintyInDays
#> 163         163 BMDE2.00-ext                 AllIndividualsReported
#> 164         164 BMDE2.00-ext                     AllSpeciesReported
#> 165         165 BMDE2.00-ext                                UTMZone
#> 166         166 BMDE2.00-ext                            UTMNorthing
#> 167         167 BMDE2.00-ext                             UTMEasting
#> 168         168 BMDE2.00-ext CoordinatesUncertaintyInDecimalDegrees
#> 169         169 BMDE2.00-ext                             CommonName
#> 170         170 BMDE2.00-ext                      RecordPermissions
#> 171         171 BMDE2.00-ext                   MultiScientificName1
#> 172         172 BMDE2.00-ext                   MultiScientificName2
#> 173         173 BMDE2.00-ext                   MultiScientificName3
#> 174         174 BMDE2.00-ext                   MultiScientificName4
#> 175         175 BMDE2.00-ext                   MultiScientificName5
#> 176         176 BMDE2.00-ext                   MultiScientificName6
#> 177         177 BMDE2.00-ext              TaxonomicAuthorityAuthors
#> 178         178 BMDE2.00-ext              TaxonomicAuthorityVersion
#> 179         179 BMDE2.00-ext                 TaxonomicAuthorityYear
#> 180         180 BMDE2.00-ext                            SpeciesCode
#> 181         181 BMDE2.00-ext                         TaxonConceptID
#> 182         182 BMDE2.00-ext                  BreedingBirdAtlasCode
#> 183         183 BMDE2.00-ext                     HabitatDescription
#> 184         184 BMDE2.00-ext                               Remarks2
#> 185         185 BMDE2.00-ext                         DateIdentified
#> 186         186 BMDE2.00-ext                           IdentifiedBy
#> 187         187 BMDE2.00-ext                  AssociatedOccurrences
#> 188         188 BMDE2.00-ext                       MeasurementType1
#> 189         189 BMDE2.00-ext                      MeasurementValue1
#> 190         190 BMDE2.00-ext                   MeasurementAccuracy1
#> 191         191 BMDE2.00-ext                       MeasurementUnit1
#> 192         192 BMDE2.00-ext               MeasurementDeterminedBy1
#> 193         193 BMDE2.00-ext                     MeasurementMethod1
#> 194         194 BMDE2.00-ext                       MeasurementType2
#> 195         195 BMDE2.00-ext                      MeasurementValue2
#> 196         196 BMDE2.00-ext                   MeasurementAccuracy2
#> 197         197 BMDE2.00-ext                       MeasurementUnit2
#> 198         198 BMDE2.00-ext               MeasurementDeterminedBy2
#> 199         199 BMDE2.00-ext                     MeasurementMethod2
#> 200         200 BMDE2.00-ext                       MeasurementType3
#> 201         201 BMDE2.00-ext                      MeasurementValue3
#> 202         202 BMDE2.00-ext                   MeasurementAccuracy3
#> 203         203 BMDE2.00-ext                       MeasurementUnit3
#> 204         204 BMDE2.00-ext               MeasurementDeterminedBy3
#> 205         205 BMDE2.00-ext                     MeasurementMethod3
#> 206         206 BMDE2.00-ext                       MeasurementType4
#> 207         207 BMDE2.00-ext                      MeasurementValue4
#> 208         208 BMDE2.00-ext                   MeasurementAccuracy4
#> 209         209 BMDE2.00-ext                       MeasurementUnit4
#> 210         210 BMDE2.00-ext               MeasurementDeterminedBy4
#> 211         211 BMDE2.00-ext                     MeasurementMethod4
#> 212         212 BMDE2.00-ext                       MeasurementType5
#> 213         213 BMDE2.00-ext                      MeasurementValue5
#> 214         214 BMDE2.00-ext                   MeasurementAccuracy5
#> 215         215 BMDE2.00-ext                       MeasurementUnit5
#> 216         216 BMDE2.00-ext               MeasurementDeterminedBy5
#> 217         217 BMDE2.00-ext                     MeasurementMethod5
#> 218         218 BMDE2.00-ext                       MeasurementType6
#> 219         219 BMDE2.00-ext                      MeasurementValue6
#> 220         220 BMDE2.00-ext                   MeasurementAccuracy6
#> 221         221 BMDE2.00-ext                       MeasurementUnit6
#> 222         222 BMDE2.00-ext               MeasurementDeterminedBy6
#> 223         223 BMDE2.00-ext                     MeasurementMethod6
#> 224         224 BMDE2.00-ext                       MeasurementType7
#> 225         225 BMDE2.00-ext                      MeasurementValue7
#> 226         226 BMDE2.00-ext                   MeasurementAccuracy7
#> 227         227 BMDE2.00-ext                       MeasurementUnit7
#> 228         228 BMDE2.00-ext               MeasurementDeterminedBy7
#> 229         229 BMDE2.00-ext                     MeasurementMethod7
#> 230         230 BMDE2.00-ext                       MeasurementType8
#> 231         231 BMDE2.00-ext                      MeasurementValue8
#> 232         232 BMDE2.00-ext                   MeasurementAccuracy8
#> 233         233 BMDE2.00-ext                       MeasurementUnit8
#> 234         234 BMDE2.00-ext               MeasurementDeterminedBy8
#> 235         235 BMDE2.00-ext                     MeasurementMethod8
#> 236         236 BMDE2.00-ext                       MeasurementType9
#> 237         237 BMDE2.00-ext                      MeasurementValue9
#> 238         238 BMDE2.00-ext                   MeasurementAccuracy9
#> 239         239 BMDE2.00-ext                       MeasurementUnit9
#> 240         240 BMDE2.00-ext               MeasurementDeterminedBy9
#> 241         241 BMDE2.00-ext                     MeasurementMethod9
#> 242         242 BMDE2.00-ext                      MeasurementType10
#> 243         243 BMDE2.00-ext                     MeasurementValue10
#> 244         244 BMDE2.00-ext                  MeasurementAccuracy10
#> 245         245 BMDE2.00-ext                      MeasurementUnit10
#> 246         246 BMDE2.00-ext              MeasurementDeterminedBy10
#> 247         247 BMDE2.00-ext                    MeasurementMethod10
#> 248         248 BMDE2.00-ext                      MeasurementType11
#> 249         249 BMDE2.00-ext                     MeasurementValue11
#> 250         250 BMDE2.00-ext                  MeasurementAccuracy11
#> 251         251 BMDE2.00-ext                      MeasurementUnit11
#> 252         252 BMDE2.00-ext              MeasurementDeterminedBy11
#> 253         253 BMDE2.00-ext                    MeasurementMethod11
#> 254         254 BMDE2.00-ext                      MeasurementType12
#> 255         255 BMDE2.00-ext                     MeasurementValue12
#> 256         256 BMDE2.00-ext                  MeasurementAccuracy12
#> 257         257 BMDE2.00-ext                      MeasurementUnit12
#> 258         258 BMDE2.00-ext              MeasurementDeterminedBy12
#> 259         259 BMDE2.00-ext                    MeasurementMethod12
#> 260         260 BMDE2.00-ext                      PrimaryMarkerType
#> 261         261 BMDE2.00-ext                    PrimaryMarkerNumber
#> 262         262 BMDE2.00-ext                    AuxiliaryMarkerType
#> 263         263 BMDE2.00-ext                  AuxiliaryMarkerNumber
#> 264         264 BMDE2.00-ext                     LastModifiedAction
#> 265         265 BMDE2.00-ext                     RecordReviewStatus
#>                                 field_name
#> 1                   GlobalUniqueIdentifier
#> 2                         DateLastModified
#> 3                            BasisOfRecord
#> 4                            InstitutionID
#> 5                          InstitutionCode
#> 6                           CollectionCode
#> 7                            CatalogNumber
#> 8                           ScientificName
#> 9                              HigherTaxon
#> 10                                 Kingdom
#> 11                                  Phylum
#> 12                                   Class
#> 13                                   Order
#> 14                                  Family
#> 15                                   Genus
#> 16                         SpecificEpithet
#> 17                       InfraspecificRank
#> 18                    InfraspecificEpithet
#> 19                    ScientificNameAuthor
#> 20                 IdentificationQualifier
#> 21                         HigherGeography
#> 22                               Continent
#> 23                               WaterBody
#> 24                             IslandGroup
#> 25                                  Island
#> 26                                 Country
#> 27                           StateProvince
#> 28                                  County
#> 29                                Locality
#> 30                MinimumElevationInMeters
#> 31                MaximumElevationInMeters
#> 32                    MinimumDepthInMeters
#> 33                    MaximumDepthInMeters
#> 34                         DecimalLatitude
#> 35                        DecimalLongitude
#> 36                           GeodeticDatum
#> 37           CoordinateUncertaintyInMeters
#> 38                           YearCollected
#> 39                          MonthCollected
#> 40                            DayCollected
#> 41                           TimeCollected
#> 42                               JulianDay
#> 43                               Collector
#> 44                                     Sex
#> 45                               LifeStage
#> 46                                ImageURL
#> 47                      RelatedInformation
#> 48                         CollectorNumber
#> 49                             FieldNumber
#> 50                              FieldNotes
#> 51               OriginalCoordinatesSystem
#> 52                         LatLongComments
#> 53                      GeoreferenceMethod
#> 54                  GeoreferenceReferences
#> 55          GeoreferenceVerificationStatus
#> 56                                 Remarks
#> 57                            FootprintWKT
#> 58                            FootprintSRS
#> 59                             ProjectCode
#> 60                            ProtocolType
#> 61                            ProtocolCode
#> 62                 ProtocolSpeciesTargeted
#> 63                       ProtocolReference
#> 64                             ProtocolURL
#> 65                    SurveyAreaIdentifier
#> 66                          SurveyAreaSize
#> 67             SurveyAreaPercentageCovered
#> 68                         SurveyAreaShape
#> 69                SurveyAreaLongAxisLength
#> 70               SurveyAreaShortAxisLength
#> 71           SurveyAreaLongAxisOrientation
#> 72                        CoordinatesScope
#> 73                 SamplingEventIdentifier
#> 74                  SamplingEventStructure
#> 75                         RouteIdentifier
#> 76                 TimeObservationsStarted
#> 77                   TimeObservationsEnded
#> 78                         DurationInHours
#> 79                     TimeIntervalStarted
#> 80                       TimeIntervalEnded
#> 81                   TimeIntervalsAdditive
#> 82                       NumberOfObservers
#> 83                      EffortMeasurement1
#> 84                            EffortUnits1
#> 85                      EffortMeasurement2
#> 86                            EffortUnits2
#> 87                      EffortMeasurement3
#> 88                            EffortUnits3
#> 89                      EffortMeasurement4
#> 90                            EffortUnits4
#> 91                      EffortMeasurement5
#> 92                            EffortUnits5
#> 93                      EffortMeasurement6
#> 94                            EffortUnits6
#> 95                      EffortMeasurement7
#> 96                            EffortUnits7
#> 97                      EffortMeasurement8
#> 98                            EffortUnits8
#> 99                      EffortMeasurement9
#> 100                           EffortUnits9
#> 101                    EffortMeasurement10
#> 102                          EffortUnits10
#> 103                    EffortMeasurement11
#> 104                          EffortUnits11
#> 105                    EffortMeasurement12
#> 106                          EffortUnits12
#> 107                    EffortMeasurement13
#> 108                          EffortUnits13
#> 109                    EffortMeasurement14
#> 110                          EffortUnits14
#> 111                    EffortMeasurement15
#> 112                          EffortUnits15
#> 113                    EffortMeasurement16
#> 114                          EffortUnits16
#> 115                    EffortMeasurement17
#> 116                          EffortUnits17
#> 117                    EffortMeasurement18
#> 118                          EffortUnits18
#> 119                         NoObservations
#> 120                   DistanceFromObserver
#> 121                DistanceFromObserverMin
#> 122                DistanceFromObserverMax
#> 123                      DistanceFromStart
#> 124                       BearingInDegrees
#> 125                SpecimenDecimalLatitude
#> 126               SpecimenDecimalLongitude
#> 127                  SpecimenGeodeticDatum
#> 128                        SpecimenUTMZone
#> 129                    SpecimenUTMNorthing
#> 130                     SpecimenUTMEasting
#> 131                       ObservationCount
#> 132                  ObservationDescriptor
#> 133                      ObservationCount2
#> 134                 ObservationDescriptor2
#> 135                      ObservationCount3
#> 136                 ObservationDescriptor3
#> 137                      ObservationCount4
#> 138                 ObservationDescriptor4
#> 139                      ObservationCount5
#> 140                 ObservationDescriptor5
#> 141                      ObservationCount6
#> 142                 ObservationDescriptor6
#> 143                      ObservationCount7
#> 144                 ObservationDescriptor7
#> 145                      ObservationCount8
#> 146                 ObservationDescriptor8
#> 147                      ObservationCount9
#> 148                 ObservationDescriptor9
#> 149                     ObservationCount10
#> 150                ObservationDescriptor10
#> 151                     ObservationCount11
#> 152                ObservationDescriptor11
#> 153                     ObservationCount12
#> 154                ObservationDescriptor12
#> 155                     ObservationCount13
#> 156                ObservationDescriptor13
#> 157                     ObservationCount14
#> 158                ObservationDescriptor14
#> 159                        ObsCountAtLeast
#> 160                         ObsCountAtMost
#> 161                        ObservationDate
#> 162                  DateUncertaintyInDays
#> 163                 AllIndividualsReported
#> 164                     AllSpeciesReported
#> 165                                UTMZone
#> 166                            UTMNorthing
#> 167                             UTMEasting
#> 168 CoordinatesUncertaintyInDecimalDegrees
#> 169                             CommonName
#> 170                      RecordPermissions
#> 171                   MultiScientificName1
#> 172                   MultiScientificName2
#> 173                   MultiScientificName3
#> 174                   MultiScientificName4
#> 175                   MultiScientificName5
#> 176                   MultiScientificName6
#> 177              TaxonomicAuthorityAuthors
#> 178              TaxonomicAuthorityVersion
#> 179                 TaxonomicAuthorityYear
#> 180                            SpeciesCode
#> 181                         TaxonConceptID
#> 182                  BreedingBirdAtlasCode
#> 183                     HabitatDescription
#> 184                               Remarks2
#> 185                         DateIdentified
#> 186                           IdentifiedBy
#> 187                  AssociatedOccurrences
#> 188                       MeasurementType1
#> 189                      MeasurementValue1
#> 190                   MeasurementAccuracy1
#> 191                       MeasurementUnit1
#> 192               MeasurementDeterminedBy1
#> 193                     MeasurementMethod1
#> 194                       MeasurementType2
#> 195                      MeasurementValue2
#> 196                   MeasurementAccuracy2
#> 197                       MeasurementUnit2
#> 198               MeasurementDeterminedBy2
#> 199                     MeasurementMethod2
#> 200                       MeasurementType3
#> 201                      MeasurementValue3
#> 202                   MeasurementAccuracy3
#> 203                       MeasurementUnit3
#> 204               MeasurementDeterminedBy3
#> 205                     MeasurementMethod3
#> 206                       MeasurementType4
#> 207                      MeasurementValue4
#> 208                   MeasurementAccuracy4
#> 209                       MeasurementUnit4
#> 210               MeasurementDeterminedBy4
#> 211                     MeasurementMethod4
#> 212                       MeasurementType5
#> 213                      MeasurementValue5
#> 214                   MeasurementAccuracy5
#> 215                       MeasurementUnit5
#> 216               MeasurementDeterminedBy5
#> 217                     MeasurementMethod5
#> 218                       MeasurementType6
#> 219                      MeasurementValue6
#> 220                   MeasurementAccuracy6
#> 221                       MeasurementUnit6
#> 222               MeasurementDeterminedBy6
#> 223                     MeasurementMethod6
#> 224                       MeasurementType7
#> 225                      MeasurementValue7
#> 226                   MeasurementAccuracy7
#> 227                       MeasurementUnit7
#> 228               MeasurementDeterminedBy7
#> 229                     MeasurementMethod7
#> 230                       MeasurementType8
#> 231                      MeasurementValue8
#> 232                   MeasurementAccuracy8
#> 233                       MeasurementUnit8
#> 234               MeasurementDeterminedBy8
#> 235                     MeasurementMethod8
#> 236                       MeasurementType9
#> 237                      MeasurementValue9
#> 238                   MeasurementAccuracy9
#> 239                       MeasurementUnit9
#> 240               MeasurementDeterminedBy9
#> 241                     MeasurementMethod9
#> 242                      MeasurementType10
#> 243                     MeasurementValue10
#> 244                  MeasurementAccuracy10
#> 245                      MeasurementUnit10
#> 246              MeasurementDeterminedBy10
#> 247                    MeasurementMethod10
#> 248                      MeasurementType11
#> 249                     MeasurementValue11
#> 250                  MeasurementAccuracy11
#> 251                      MeasurementUnit11
#> 252              MeasurementDeterminedBy11
#> 253                    MeasurementMethod11
#> 254                      MeasurementType12
#> 255                     MeasurementValue12
#> 256                  MeasurementAccuracy12
#> 257                      MeasurementUnit12
#> 258              MeasurementDeterminedBy12
#> 259                    MeasurementMethod12
#> 260                      PrimaryMarkerType
#> 261                    PrimaryMarkerNumber
#> 262                    AuxiliaryMarkerType
#> 263                  AuxiliaryMarkerNumber
#> 264                     LastModifiedAction
#> 265                     RecordReviewStatus