Access and download data on plant and animal populations from NatureCounts

naturecounts is an R package for accessing and downloading data on plant and animal populations from various databases through NatureCounts, a service managed by Bird Studies Canada.

Details

See the vignettes (vignettes(package = "naturecounts"), or at https://birdscanada.github.io/naturecounts for getting started!

See also

Author

Maintainer: Denis Lepage dlepage@birdscanada.org

Authors:

Steffi LaZerte steffi@steffi.ca