Download information about NatureCounts collections

Download the number of records available for different collections filtered by location (if provided). If authorization is provided, the collections are filtered to only those available to the user (unless using show = "all"). Without authorization all collections are returned.

nc_count(
  collections = NULL,
  project_ids = NULL,
  species = NULL,
  years = NULL,
  doy = NULL,
  region = NULL,
  site_type = NULL,
  show = "available",
  username = NULL,
  timeout = 120,
  verbose = TRUE
)

Arguments

collections
Character vector. The collection codes from which to download data. NULL (default) downloads data from all available collections

project_ids
Character/Numeric vector. The project ids from which to download data. First the collections associated with a project_id are determined, and then data is downloaded for each collection. If both collections and project_ids are supplied, they are combined.

species
Numeric vector. Numeric species ids (see details)

years
Numeric vector. The start/end years of data to download. Can use NA for either start or end, or a single value to return data from a single year.

doy
Character/Numeric vector. The start/end day-of-year to download (1-366 or dates that can be converted to day of year). Can use NA for either start or end

region
List. Named list with one of the following options: country, statprov, subnational2, iba, bcr, utm_squares, bbox. See details

site_type
Character vector. The type of site to return (e.g., IBA).

show
Character. Either "all" or "available". "all" returns counts from all data sources. "available" only returns counts for data available for the username provided. If no username is provided, defaults to "all".

username
Character vector. Username for http://naturecounts.ca. If provided, the user will be prompted for a password. If left NULL, only public collections will be returned.

timeout
Numeric. Number of seconds before connecting to the server times out.

verbose
Logical. Show messages?

Value

Data frame

Details

The akn_level column describes the level of data access for that collection (see descriptions online). The access column describes the accessibility of a collection for a given username (or no access if no username supplied). See the section on Access and request_ids for more details.

NatureCounts account

All public data is available with a username/password (sign up for a free NatureCounts account). However, to access private/semi-public projects/collections you must request access. See the Access and request_ids section for more information.

Species ids (species)

Numeric species id codes can determined from the functions search_species() or search_species_code(). See also the article on species codes for more information.

Day of Year (doy)

The format for day of year (doy) is fairly flexible and can be a whole number between 1 and 366 or anything recognized by lubridate-package's ymd() function. However, it must have the order of year, month, day. Note that year is ignored when converting to day of year, except that it will result in a 1 day offset for leap years.

Regions (region)

Regions are defined by codes reflecting the country, state/province, subnational (level 2), Important Bird Areas (IBA), and Bird Conservation Regions (BCR) (see search_region() for codes). They can also be defined by providing specific UTM squares to download or a bounding box area which specifies the min/max longitude and min/max latitude (bbox). See the article on regional filters for more information.

Access and request_ids

Access to a data collection is either available as "full" or "by request". Use nc_count(username = "USER", show = "all"), to see the accessibility of collections.

"Full" access means that data can be immediately requested directly through the naturecounts R package. "By request" means that a request must be submitted online and approved before the data can be downloaded through naturecounts.

This means that there are two types of data requests: ones made through this naturecounts R package (API requests) and those made through the online Web Request Form (Web requests). Every request (from either method) generates a request_id which identifies the filter set and collections requested. Details of all of requests can be reviewed with the nc_requests() function.

To download data with "full" access, users can either specify filters, or if they are repeating a download, can use the request_id from nc_requests().

Otherwise, if the user doesn't have "full" access, they must supply an approved request_id to the nc_data_dl() function (e.g., nc_data_dl(request_id = 152000, username = "USER")). Use nc_requests() to see request_ids, filters, and approval status.

Requests for "full" access to additional collections can be made online through the Web Request Form by checking the "Full access?" box in Step 2 of the form.

See also

Examples

# Count all publicly available records:
# \donttest{
nc_count()

#> Without a username, using 'show = "all"'

#>                collection akn_level     access nrecords
#> 1                ABATLAS1         5       full   123364
#> 2                ABATLAS2         5       full   201398
#> 3              ABBIRDRECS         5       full   357264
#> 4                  ABOWLS         4 by request        0
#> 5                   ATBBS         4 by request        0
#> 6                  ATOWLS         4 by request    25192
#> 7                     BBS         5       full  5735895
#> 8                   BBS50         5       full        0
#> 9               BBS50-CAN         5       full  4168740
#> 10              BBS50-MEX         2  no access        0
#> 11          BBS50-US-EAST         5       full  9940162
#> 12          BBS50-US-WEST         5       full 10749161
#> 13          BCATLAS1BE_DO         5       full    41412
#> 14         BCATLAS1BE_RAW         5       full   344433
#> 15        BCATLAS1BE_SUMM         5       full   165485
#> 16             BCATLAS1PC         5       full   260647
#> 17         BCATLAS1PC_SAR         3 by request        0
#> 18             BCATLAS1RC         3 by request    21884
#> 19                BCBEACH         4 by request    10182
#> 20                  BCCWS         4 by request   353868
#> 21                   BCMA         5       full    11291
#> 22                 BCOWLS         4 by request    28577
#> 23                  BGRMM         5       full      476
#> 24              BIRDSOFBC         2  no access   498444
#> 25                    CBC         2  no access  7570427
#> 26                   CLLS         4 by request        0
#> 27         CMMN-BAND-LPBO         3 by request  1168735
#> 28           CMMN-DET-ABO         3 by request    85569
#> 29          CMMN-DET-ACBO         4 by request    28386
#> 30          CMMN-DET-ATBP         2  no access        0
#> 31          CMMN-DET-ATSI         2  no access        0
#> 32           CMMN-DET-BBO         3 by request    72759
#> 33        CMMN-DET-BIBMRS         2  no access        0
#> 34          CMMN-DET-BPBO         3 by request    92639
#> 35          CMMN-DET-DMBO         2  no access    88280
#> 36          CMMN-DET-GMNP         2  no access        0
#> 37           CMMN-DET-HBO         5       full   274872
#> 38           CMMN-DET-IBS         3 by request    13607
#> 39          CMMN-DET-IPBO         3 by request    51964
#> 40          CMMN-DET-LMBO         3 by request    57432
#> 41          CMMN-DET-LPBO         3 by request  1101083
#> 42         CMMN-DET-LSLBO         3 by request   114040
#> 43          CMMN-DET-MGBO         3 by request    94662
#> 44          CMMN-DET-MIBO         4 by request    12180
#> 45           CMMN-DET-MNO         3 by request    69732
#> 46           CMMN-DET-OOT         3 by request   114793
#> 47         CMMN-DET-PEPBO         3 by request   144454
#> 48          CMMN-DET-PIBO         3 by request   124846
#> 49          CMMN-DET-RBMS         2  no access        0
#> 50          CMMN-DET-ROCK         3 by request        0
#> 51          CMMN-DET-RPBO         2  no access   144048
#> 52          CMMN-DET-RUTH         5       full        0
#> 53         CMMN-DET-SABBS         2  no access        0
#> 54          CMMN-DET-SELK         5       full        0
#> 55          CMMN-DET-TCBO         3 by request   136565
#> 56         CMMN-DET-TLBBS         4 by request    33522
#> 57          CMMN-DET-TLBO         3 by request    29722
#> 58        CMMN-DET-TTPBRS         3 by request   112862
#> 59          CMMN-DET-VLBO         5       full    56287
#> 60          CMMN-DET-WPBO         2  no access    63251
#> 61         CMMN-LPBO-MOBU         3 by request     2970
#> 62         CMMN-PIBO-MOBU         3 by request        0
#> 63            EBIRD-CA-AT         4 by request  4811070
#> 64            EBIRD-CA-BC         4 by request 13496087
#> 65            EBIRD-CA-NO         4 by request   880775
#> 66            EBIRD-CA-ON         4 by request 24750269
#> 67            EBIRD-CA-PR         4 by request  7461849
#> 68            EBIRD-CA-QC         4 by request 19555444
#> 69          EBIRD-CA-SENS         3 by request    33272
#> 70             EBUTTERFLY         5       full    26192
#> 71                   EPOQ         2  no access        0
#> 72                   GBBC         4 by request   431613
#> 73                GBHE_BC         3 by request        0
#> 74              HawkCount         4 by request   112673
#> 75                   HELP         3 by request        0
#> 76                   IMMP         5       full       43
#> 77                IMMP_A2         5       full       24
#> 78                IMMP_BW         5       full       19
#> 79                   IMQC         5       full      761
#> 80            INATURALIST         3 by request        0
#> 81              LPWCENSUS         2  no access        0
#> 82          MBATLAS1BE_DO         5       full    93876
#> 83         MBATLAS1BE_RAW         5       full   326429
#> 84        MBATLAS1BE_SUMM         5       full   154856
#> 85             MBATLAS1PC         5       full   306094
#> 86         MBATLAS1PC_SAR         3 by request        0
#> 87             MBATLAS1RC         3 by request    15622
#> 88            MBBA1BE_RAW         5       full   142920
#> 89           MBBA1BE_SUMM         5       full    93053
#> 90            MBBA2BE_RAW         5       full   263809
#> 91           MBBA2BE_SUMM         5       full   114616
#> 92                MBBA2PC         5       full   127447
#> 93            MBBA2PC_SAR         3 by request       13
#> 94                MBBA2RC         3 by request    14213
#> 95                 MBOWLS         4 by request    16382
#> 96                   MEXU         5       full     3251
#> 97                   MLMP         5       full      802
#> 98                     MM         5       full     4343
#> 99                   MMMP         3 by request    88737
#> 100              MMPBIRDS         4 by request   393302
#> 101              MMPFROGS         4 by request    91700
#> 102                  MNRS         4 by request        0
#> 103          MONARCHWATCH         5       full   145665
#> 104           NATURALISTA         5       full       30
#> 105                NBFHWS         3 by request        0
#> 106             NESTWATCH         3 by request   413415
#> 107   NESTWATCH_ACAMFIELD         3 by request        0
#> 108     NESTWATCH_KMARTIN         5       full        0
#> 109  NESTWATCH_TLEUKERING         2  no access        0
#> 110         NFATLAS1BE_DO         4 by request      243
#> 111       NFATLAS1BE_SUMM         4 by request      180
#> 112            NFATLAS1PC         4 by request        0
#> 113            NFATLAS1RC         4 by request        0
#> 114        NRCS_ACAMFIELD         3 by request        0
#> 115          NRCS_KMARTIN         5       full        0
#> 116           OBBA1BE_RAW         5       full   407287
#> 117          OBBA1BE_SUMM         5       full   205816
#> 118           OBBA2BE_RAW         5       full   541743
#> 119          OBBA2BE_SUMM         5       full   256702
#> 120               OBBA2PC         5       full   592425
#> 121           OBBA2PC_SAR         3 by request       31
#> 122               OBBA2RC         3 by request    27970
#> 123                  OBFS         5       full   324560
#> 124                   OHI         4 by request        0
#> 125             OISEAUXQC         2  no access   666963
#> 126                ONOWLS         4 by request    34940
#> 127                  ONRS         4 by request        0
#> 128                   PFW         5       full  3841194
#> 129           PFW-US-EAST         4 by request  8651432
#> 130           PFW-US-WEST         4 by request  5631851
#> 131                  PMMM         5       full        5
#> 132                  PNRS         4 by request        0
#> 133             PPNP-FBMP         4 by request        0
#> 134             PPNP-MOBU         3 by request        0
#> 135            PRISM-ACSS         4 by request   123002
#> 136             PRISM-OSS         4 by request    48139
#> 137             PRISM-PSS         3 by request     4287
#> 138   QCATLAS_NORTH_BE_DO         3 by request    20430
#> 139  QCATLAS_NORTH_BE_RAW         3 by request    39963
#> 140 QCATLAS_NORTH_BE_SUMM         3 by request    25816
#> 141      QCATLAS_NORTH_PC         3 by request    24324
#> 142        QCATLAS1BE_RAW         5       full   198816
#> 143       QCATLAS1BE_SUMM         5       full   136333
#> 144         QCATLAS2BE_DO         5       full   283034
#> 145        QCATLAS2BE_RAW         5       full   601540
#> 146       QCATLAS2BE_SUMM         5       full   231775
#> 147            QCATLAS2PC         5       full   341902
#> 148        QCATLAS2PC_SAR         2  no access        0
#> 149            QCATLAS2RC         2  no access    39390
#> 150                 QCBBS         4 by request        0
#> 151                QCOWLS         4 by request    12820
#> 152                  QNRS         4 by request        0
#> 153           RCBIOTABASE         5       full    12614
#> 154              SAR-ACFL         1  no access       NA
#> 155         SAR-BAEA-NEST         1  no access       NA
#> 156        SAR-BAEA-SATEL         1  no access       NA
#> 157              SAR-BNOW         1  no access       NA
#> 158              SAR-GWWA         1  no access       NA
#> 159              SAR-HOWA         1  no access       NA
#> 160            SAR-NSPIPL         1  no access       NA
#> 161              SAR-PROW         1  no access       NA
#> 162              SAR-RSHA         3 by request    27696
#> 163              SAR-SEOW         1  no access       NA
#> 164         SKATLAS1BE_DO         4 by request   270718
#> 165       SKATLAS1BE_SUMM         4 by request    94929
#> 166            SKATLAS1PC         4 by request   118190
#> 167            SKATLAS1RC         4 by request        0
#> 168                SKOWLS         3 by request     2863
#> 169            SWIFTWATCH         3 by request    12295
#> 170                   VBS         5 by request        0
#> 171                  WMMM         5       full     8981
#> 172                  WPWI         5       full     3012
#> 173            WRNIGHTJAR         3 by request     6368

# }

# Count publicly available records for Manitoba, Canada
# \donttest{
nc_count(region = list(statprov = "MB"))

#> Without a username, using 'show = "all"'

#> Using filters: statprov (MB)

#>         collection akn_level     access nrecords
#> 1       ABBIRDRECS         5       full      357
#> 2              BBS         5       full    86997
#> 3        BBS50-CAN         5       full   453998
#> 4              CBC         2  no access    20507
#> 5    CMMN-DET-DMBO         2  no access    88280
#> 6      EBIRD-CA-PR         4 by request  2070346
#> 7    EBIRD-CA-SENS         3 by request     3510
#> 8       EBUTTERFLY         5       full      224
#> 9             GBBC         4 by request    24393
#> 10       HawkCount         4 by request     1446
#> 11            IMQC         5       full       26
#> 12   MBATLAS1BE_DO         5       full    93876
#> 13  MBATLAS1BE_RAW         5       full   326429
#> 14 MBATLAS1BE_SUMM         5       full   154856
#> 15      MBATLAS1PC         5       full   306094
#> 16      MBATLAS1RC         3 by request    15622
#> 17          MBOWLS         4 by request    15314
#> 18              MM         5       full       53
#> 19        MMPBIRDS         4 by request       60
#> 20    MONARCHWATCH         5       full       24
#> 21       NESTWATCH         3 by request     4553
#> 22       OISEAUXQC         2  no access        2
#> 23             PFW         5       full   110414
#> 24   SKATLAS1BE_DO         4 by request        1

# }

# Count all records for all collections user "sample" has access to
if (FALSE) {
nc_count(username = "sample")
}

# Count records with barred owls in Ontario
search_species("barred owl")

#>    species_id                            scientific_name
#> 1        7590                                Strix varia
#> 2        7595                 Strix occidentalis x varia
#> 3       32704                            Strix hylophila
#> 4       32746                      Glaucidium cuculoides
#> 5       32750                         Glaucidium capense
#> 6       32805                       Aegotheles bennettii
#> 7       43347              Glaucidium capense scheffleri
#> 8       43348       Glaucidium capense capense/ngamiense
#> 9       43362                  Strix varia [varia Group]
#> 10      43363                       Strix varia sartorii
#> 11      43377 Aegotheles bennettii bennettii/wiedenfeldi
#> 12      43378             Aegotheles bennettii terborghi
#> 13      43379              Aegotheles bennettii plumifer
#> 14      45499                       Aegotheles bennettii
#> 15      47281                   Strix occidentalis/varia
#>                           english_name
#> 1                           Barred Owl
#> 2        Spotted x Barred Owl (hybrid)
#> 3                     Rusty-barred Owl
#> 4                   Asian Barred Owlet
#> 5                 African Barred Owlet
#> 6                Barred Owlet-Nightjar
#> 7  African Barred Owlet (Spot-fronted)
#> 8   African Barred Owlet (Bar-fronted)
#> 9                Barred Owl (Northern)
#> 10              Barred Owl (Cinereous)
#> 11      Barred Owlet-nightjar (Barred)
#> 12   Barred Owlet-nightjar (Diamond's)
#> 13       Barred Owlet-nightjar (Dwarf)
#> 14               Barred Owlet-nightjar
#> 15                  Spotted/Barred Owl
#>                                       french_name taxon_group
#> 1                                  Chouette rayée       BIRDS
#> 2            Hybride Chouette tachetée x C. rayée       BIRDS
#> 3                                 Chouette dryade       BIRDS
#> 4                           Chevêchette cuculoïde       BIRDS
#> 5                              Chevêchette du Cap       BIRDS
#> 6                             Égothèle de Bennett       BIRDS
#> 7                 Chevêchette du Cap (scheffleri)       BIRDS
#> 8       Chevêchette du Cap [capense ou ngamiense]       BIRDS
#> 9                          Chouette rayée [varia]       BIRDS
#> 10                      Chouette rayée (sartorii)       BIRDS
#> 11 Égothèle de Bennett [bennettii ou wiedenfeldi]       BIRDS
#> 12                Égothèle de Bennett (terborghi)       BIRDS
#> 13                 Égothèle de Bennett (plumifer)       BIRDS
#> 14                            Égothèle de Bennett       BIRDS
#> 15                  Chouette tachetée ou C. rayée       BIRDS

nc_count(species = 7590, region = list(statprov = "ON"), username = "sample")

#> Using filters: species (7590); statprov (ON)

#>   collection akn_level access nrecords
#> 1    SAMPLE1         0   full       NA
#> 2    SAMPLE2         0   full        1

# Count all records available in the Christmas Bird Count and Breeding Bird
# Survey collections (regardless of user permissions)
nc_count(collections = c("CBC", "BBS"), show = "all", username = "sample")

#> Using filters: collections (CBC, BBS)

#>   collection akn_level     access nrecords
#> 1        BBS         5 by request  5735895
#> 2        CBC         2  no access  7570427