Find species codes

Find species id codes by searching for scientific, English and French species names.

search_species(name = NULL, show = "names", authority = NULL)

Arguments

name

Character. The species name to search for

show

Character. Either "all" or "names" (default). Whether to return all taxonomic information or only a subset with species names

authority

Character. If not NULL (default), return the alphanumeric code associated with avian species for this taxonomic authority.

Value

Data frame of species ids and taxonomic information

Details

species_search() is deprecated in favour of search_species()

Examples

# Show all ids
search_species()
#> # A tibble: 37,104 × 5
#>    species_id scientific_name          english_name              frenc…¹ taxon…²
#>         <int> <chr>                    <chr>                     <chr>   <chr>  
#>  1          0 N/A                      No observations           Aucune… BIRDS  
#>  2         10 Tinamus major            Great Tinamou             Grand … BIRDS  
#>  3         20 Nothocercus bonapartei   Highland Tinamou          Tinamo… BIRDS  
#>  4         30 Crypturellus soui        Little Tinamou            Tinamo… BIRDS  
#>  5         40 Crypturellus cinnamomeus Thicket Tinamou           Tinamo… BIRDS  
#>  6         50 Crypturellus boucardi    Slaty-breasted Tinamou    Tinamo… BIRDS  
#>  7         60 Crypturellus kerriae     Choco Tinamou             Tinamo… BIRDS  
#>  8         70 Dendrocygna viduata      White-faced Whistling-Du… Dendro… BIRDS  
#>  9         80 Dendrocygna autumnalis   Black-bellied Whistling-… Dendro… BIRDS  
#> 10         90 Dendrocygna arborea      West Indian Whistling-Du… Dendro… BIRDS  
#> # … with 37,094 more rows, and abbreviated variable names ¹​french_name,
#> #   ²​taxon_group

search_species("chickadee")
#> # A tibble: 20 × 5
#>    species_id scientific_name                           englis…¹ frenc…² taxon…³
#>         <int> <chr>                                     <chr>    <chr>   <chr>  
#>  1      14270 Poecile carolinensis                      Carolin… Mésang… BIRDS  
#>  2      14280 Poecile atricapillus                      Black-c… Mésang… BIRDS  
#>  3      14290 Poecile gambeli                           Mountai… Mésang… BIRDS  
#>  4      14300 Poecile sclateri                          Mexican… Mésang… BIRDS  
#>  5      14310 Poecile rufescens                         Chestnu… Mésang… BIRDS  
#>  6      14320 Poecile hudsonicus                        Boreal … Mésang… BIRDS  
#>  7      14330 Poecile cinctus                           Gray-he… Mésang… BIRDS  
#>  8      14388 Paridae sp.                               Chickad… Mésang… BIRDS  
#>  9      40668 Poecile carolinensis x atricapillus       Carolin… Hybrid… BIRDS  
#> 10      40669 Poecile carolinensis/atricapillus         Carolin… Mésang… BIRDS  
#> 11      40670 Poecile gambeli [gambeli Group]           Mountai… Mésang… BIRDS  
#> 12      40671 Poecile gambeli [baileyae Group]          Mountai… Mésang… BIRDS  
#> 13      40932 Poecile atricapillus x gambeli            Black-c… Hybrid… BIRDS  
#> 14      40955 Poecile sp.                               chickad… Poecil… BIRDS  
#> 15      41990 Poecile gambeli x hudsonicus              Mountai… Hybrid… BIRDS  
#> 16      44462 Poecile atricapillus x hudsonicus         Black-c… Hybrid… BIRDS  
#> 17      44463 Poecile montanus x cinctus                Willow … Hybrid… BIRDS  
#> 18      44465 Poecile carolinensis x Baeolophus bicolor Carolin… Hybrid… BIRDS  
#> 19      44466 Poecile atricapillus x Baeolophus bicolor Black-c… Hybrid… BIRDS  
#> 20      45746 Poecile atricapillus/gambeli              Black-c… Mésang… BIRDS  
#> # … with abbreviated variable names ¹​english_name, ²​french_name, ³​taxon_group
search_species("black-capped chickadee")
#> # A tibble: 4 × 5
#>   species_id scientific_name                           english…¹ frenc…² taxon…³
#>        <int> <chr>                                     <chr>     <chr>   <chr>  
#> 1      14280 Poecile atricapillus                      Black-ca… Mésang… BIRDS  
#> 2      40668 Poecile carolinensis x atricapillus       Carolina… Hybrid… BIRDS  
#> 3      40669 Poecile carolinensis/atricapillus         Carolina… Mésang… BIRDS  
#> 4      44466 Poecile atricapillus x Baeolophus bicolor Black-ca… Hybrid… BIRDS  
#> # … with abbreviated variable names ¹​english_name, ²​french_name, ³​taxon_group

# Add alphanumeric code for BSCDATA authority
search_species("black-capped chickadee", authority = "BSCDATA")
#> # A tibble: 29 × 6
#>    species_id scientific_name      english_name          frenc…¹ taxon…² BSCDATA
#>         <int> <chr>                <chr>                 <chr>   <chr>   <chr>  
#>  1      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#>  2      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#>  3      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#>  4      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   07350  
#>  5      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#>  6      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#>  7      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   7149   
#>  8      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#>  9      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   8868   
#> 10      14280 Poecile atricapillus Black-capped Chickad… Mésang… BIRDS   BCCH   
#> # … with 19 more rows, and abbreviated variable names ¹​french_name,
#> #   ²​taxon_group

# Show all taxonomic information
search_species("black-capped chickadee", show = "all")
#> # A tibble: 4 × 17
#>   speci…¹ conce…² scien…³ speci…⁴ added…⁵ engli…⁶ frenc…⁷ phylum famil…⁸ group…⁹
#>     <int> <chr>   <chr>     <int> <chr>   <chr>   <chr>   <chr>  <chr>     <int>
#> 1   14280 B1F0CC… Poecil…       1 NA      Black-… Mésang… Chord… NA          178
#> 2   40668 1B9B3D… Poecil…       4 NA      Caroli… Hybrid… Chord… NA          178
#> 3   40669 F648AB… Poecil…       2 NA      Caroli… Mésang… Chord… NA          178
#> 4   44466 6D7E1C… Poecil…       4 NA      Black-… Hybrid… Chord… NA          178
#> # … with 7 more variables: order_taxon <chr>, taxon_group <chr>,
#> #   sort_order <int>, class <chr>, family_name <chr>, concept_source <chr>,
#> #   family_english_name <chr>, and abbreviated variable names ¹​species_id,
#> #   ²​concept_id, ³​scientific_name, ⁴​species_status, ⁵​added_dt, ⁶​english_name,
#> #   ⁷​french_name, ⁸​family_french_name, ⁹​group_id

# Using the codes
nc_count(species = 14280)
#> Without a username, using 'show = "all"'
#> Using filters: species (14280)
#> # A tibble: 93 × 4
#>    collection      akn_level access nrecords
#>    <chr>               <int> <chr>     <int>
#>  1 ABATLAS1                5 full       1314
#>  2 ABATLAS2                5 full       3773
#>  3 ABBIRDRECS              5 full       2931
#>  4 BBS                     5 full      39794
#>  5 BBS50-CAN               5 full      48423
#>  6 BBS50-US-EAST           5 full      58161
#>  7 BBS50-US-WEST           5 full      51535
#>  8 BCATLAS1BE_DO           5 full        447
#>  9 BCATLAS1BE_RAW          5 full       3616
#> 10 BCATLAS1BE_SUMM         5 full       1553
#> # … with 83 more rows