Search for bird species id codes by alphanumeric codes

This is an advanced function for returning all Bird-related species id codes based on the various alphanumeric codes used by different authorities.

search_species_code(code = NULL, authority = "BSCDATA", results = "all")

Arguments

code
Vector. Character or numeric code indicating a species for a given authority.

authority
Character. The authority to compare codes against (defaults to "BSCDATA")

results
Character. "all" returns codes for all related species (including subspecies and main species). "exact" returns only the code for exact species indicated by the code.

Value

A data frame of numeric species id codes and names

Details

species_code_search() is deprecated in favour of search_species_code()

Species ids returned reflect both species and sub-species levels.

Examples

# Show all ids
search_species_code()

#> # A tibble: 1,127 x 5
#>    species_id BSCDATA scientific_name     english_name       french_name        
#>         <int> <chr>   <chr>               <chr>              <chr>              
#>  1         80 BBWD    Dendrocygna autumn… Black-bellied Whi… Dendrocygne à vent…
#>  2        100 FUWD    Dendrocygna bicolor Fulvous Whistling… Dendrocygne fauve  
#>  3        110 BEGO    Anser fabalis/serr… Taiga/Tundra Bean… Oie des moissons o…
#>  4        111 TABG    Anser fabalis       Taiga Bean-Goose   Oie des moissons   
#>  5        112 TUBG    Anser serrirostris  Tundra Bean-Goose  Oie de la toundra  
#>  6        120 PFGO    Anser brachyrhynch… Pink-footed Goose  Oie à bec court    
#>  7        130 GWFG    Anser albifrons     Greater White-fro… Oie rieuse         
#>  8        140 LWFG    Anser erythropus    Lesser White-fron… Oie naine          
#>  9        141 GRGO    Anser anser         Graylag Goose      Oie cendrée        
#> 10        142 BHGO    Anser indicus       Bar-headed Goose   Oie à tête barrée  
#> # … with 1,117 more rows

# Get all species ids for house finches
search_species_code("HOFI")

#> # A tibble: 1 x 5
#>   species_id BSCDATA scientific_name      english_name french_name     
#>        <int> <chr>   <chr>                <chr>        <chr>           
#> 1      20350 HOFI    Haemorhous mexicanus House Finch  Roselin familier

# Get all species ids for Dark-eyed Juncos
search_species_code("DEJU")

#> # A tibble: 4 x 5
#>   species_id BSCDATA scientific_name       english_name       french_name       
#>        <int> <chr>   <chr>                 <chr>              <chr>             
#> 1      19090 SCJU    Junco hyemalis hyema… Dark-eyed Junco (… Junco ardoisé [hy…
#> 2      19100 ORJU    Junco hyemalis [oreg… Dark-eyed Junco (… Junco ardoisé [or…
#> 3      19110 PSJU    Junco hyemalis mearn… Dark-eyed Junco (… Junco ardoisé (me…
#> 4      42218 DEJU    Junco hyemalis        Dark-eyed Junco    Junco ardoisé     

# Get all species ids related to Yellow-rumped Warbler (Myrtle)
# NOTE! This includes Audubon's and the main, Yellow-rumped Warbler species
search_species_code("MYWA")

#> # A tibble: 3 x 5
#>   species_id BSCDATA scientific_name     english_name        french_name        
#>        <int> <chr>   <chr>               <chr>               <chr>              
#> 1      16610 YRWA    Setophaga coronata  Yellow-rumped Warb… Paruline à croupio…
#> 2      16620 MYWA    Setophaga coronata… Yellow-rumped Warb… Paruline à croupio…
#> 3      16630 AUWA    Setophaga coronata… Yellow-rumped Warb… Paruline à croupio…

# Get ONLY specific id related to Yellow-rumped Warbler (Myrtle)
search_species_code("MYWA", results = "exact")

#> # A tibble: 1 x 5
#>   species_id BSCDATA scientific_name      english_name       french_name        
#>        <int> <chr>   <chr>                <chr>              <chr>              
#> 1      16620 MYWA    Setophaga coronata … Yellow-rumped War… Paruline à croupio…

# Use the Christmas Bird Count authority
search_species_code(11609, authority = "CBC")

#> # A tibble: 6 x 5
#>   species_id CBC   scientific_name       english_name        french_name        
#>        <int> <chr> <chr>                 <chr>               <chr>              
#> 1      16610 11609 Setophaga coronata    Yellow-rumped Warb… Paruline à croupio…
#> 2      16620 11610 Setophaga coronata c… Yellow-rumped Warb… Paruline à croupio…
#> 3      16630 11611 Setophaga coronata a… Yellow-rumped Warb… Paruline à croupio…
#> 4      40770 11612 Setophaga coronata c… Yellow-rumped Warb… Paruline à croupio…
#> 5      40771 11613 Setophaga coronata n… Yellow-rumped Warb… Paruline à croupio…
#> 6      40772 11614 Setophaga coronata g… Yellow-rumped Warb… Paruline à croupio…

# Look in more than one authority (note that the code only needs to match on
# of the authorities)
search_species_code("MYWA", authority = c("BCMA", "CBC"))

#> # A tibble: 6 x 6
#>   species_id BCMA  CBC   scientific_name     english_name      french_name      
#>        <int> <chr> <chr> <chr>               <chr>             <chr>            
#> 1      16630 AUWA  11611 Setophaga coronata… Yellow-rumped Wa… Paruline à croup…
#> 2      16620 MYWA  11610 Setophaga coronata… Yellow-rumped Wa… Paruline à croup…
#> 3      16610 YRWA  11609 Setophaga coronata  Yellow-rumped Wa… Paruline à croup…
#> 4      40770 NA    11612 Setophaga coronata… Yellow-rumped Wa… Paruline à croup…
#> 5      40771 NA    11613 Setophaga coronata… Yellow-rumped Wa… Paruline à croup…
#> 6      40772 NA    11614 Setophaga coronata… Yellow-rumped Wa… Paruline à croup…