Zero-fill data

Zero-fill data

format_zero_fill(df_db, by = "event", species = "all",
  fill = "ObservationCount", extra_cols = NULL, verbose = TRUE)

Arguments

df_db
Either data frame or a connection to database with naturecounts table

by
Character vector. Either "event" or a vector of specific column names to fill by (see details)

species
Character vector. Either "all", for species in the data, or a vector of species ID codes to fill in.

fill
Character. The column name to fill in. Defaults to "ObservationCount".

extra_cols
Character vector. Extra columns/fields uniquely associated with by to keep in the data (all columns not in by, species, fill, or extra_cols will be omitted from the result.

verbose
Logical. Show messages?

Value

Data frame

Details

by refers to the combination of columns which are used to detect missing values. When by = "event", project_id, collection, date, and SamplingEventIdentifier are used. Otherwise users can specify their own combination of columns.

If species is supplied, all records will be used to determine observation events, but only records (zero-filled or otherwise) which correspond to a species in species will be returned (all others will be omitted). Note that records where species_id is NA (generally for 0 counts for presence/absence), will be converted to a list of 0's for the individual species.

Examples

# Download data
rc <- nc_data_dl(collection = "RCBIOTABASE", username = "sample")

#> Using filters: collections (RCBIOTABASE); fields_set (BMDE2.00-min)

#> Collecting available records...

#>    collection nrecords
#> 1 RCBIOTABASE    12598

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 5000 / 12598

#>     Records 5001 to 10000 / 12598

#>     Records 10001 to 12598 / 12598

# Remove casual observations (i.e. 'AllSpeciesReported' = "No")
library(dplyr) # For filter function

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

rc <- filter(rc, AllSpeciesReported == "Yes")

# Add dates
rc_dates <- format_dates(rc)

# Zero fill by all species present
rc_all_zeros <- format_zero_fill(rc_dates)

#> Converted 'fill' column (ObservationCount) from character to numeric

# Zero fill only for Canada Goose
rc_goose <- format_zero_fill(rc_dates, species = "230")

#> Converted 'fill' column (ObservationCount) from character to numeric