Download information about NatureCounts collections

Download the number of records available for different collections filtered by location (if provided). If authorization is provided, the collections are filtered to only those available to the user (unless using show = "all"). Without authorization all collections are returned.

nc_count(
  collections = NULL,
  project_ids = NULL,
  species = NULL,
  years = NULL,
  doy = NULL,
  region = NULL,
  site_type = NULL,
  show = "available",
  username = NULL,
  timeout = 120,
  verbose = TRUE
)

Arguments

collections
Character vector. The collection codes from which to download data. NULL (default) downloads data from all available collections

project_ids
Character/Numeric vector. The project ids from which to download data. First the collections associated with a project_id are determined, and then data is downloaded for each collection. If both collections and project_ids are supplied, they are combined.

species
Numeric vector. Numeric species ids (see details)

years
Numeric vector. The start/end years of data to download. Can use NA for either start or end, or a single value to return data from a single year.

doy
Character/Numeric vector. The start/end day-of-year to download (1-366 or dates that can be converted to day of year). Can use NA for either start or end

region
List. Named list with one of the following options: country, statprov, subnational2, iba, bcr, utm_squares, bbox. See details

site_type
Character vector. The type of site to return (e.g., IBA).

show
Character. Either "all" or "available". "all" returns counts from all data sources. "available" only returns counts for data available for the username provided. If no username is provided, defaults to "all".

username
Character vector. Username for http://naturecounts.ca. If provided, the user will be prompted for a password. If left NULL, only public collections will be returned.

timeout
Numeric. Number of seconds before connecting to the server times out.

verbose
Logical. Show messages?

Value

Data frame

Details

The akn_level column describes the level of data access for that collection (see descriptions online). The access column describes the accessibility of a collection for a given username (or no access if no username supplied). See the section on Access and request_ids for more details.

NatureCounts account

All public data is available with a username/password (sign up for a free NatureCounts account). However, to access private/semi-public projects/collections you must request access. See the Access and request_ids section for more information.

Species ids (species)

Numeric species id codes can determined from the functions search_species() or search_species_code(). See also the article on species codes for more information.

Day of Year (doy)

The format for day of year (doy) is fairly flexible and can be a whole number between 1 and 366 or anything recognized by lubridate-package's ymd() function. However, it must have the order of year, month, day. Note that year is ignored when converting to day of year, except that it will result in a 1 day offset for leap years.

Regions (region)

Regions are defined by codes reflecting the country, state/province, subnational (level 2), Important Bird Areas (IBA), and Bird Conservation Regions (BCR) (see search_region() for codes). They can also be defined by providing specific UTM squares to download or a bounding box area which specifies the min/max longitude and min/max latitude (bbox). See the article on regional filters for more information.

Access and request_ids

Access to a data collection is either available as "full" or "by request". Use nc_count(username = "USER", show = "all"), to see the accessibility of collections.

"Full" access means that data can be immediately requested directly through the naturecounts R package. "By request" means that a request must be submitted online and approved before the data can be downloaded through naturecounts.

This means that there are two types of data requests: ones made through this naturecounts R package (API requests) and those made through the online Web Request Form (Web requests). Every request (from either method) generates a request_id which identifies the filter set and collections requested. Details of all of requests can be reviewed with the nc_requests() function.

To download data with "full" access, users can either specify filters, or if they are repeating a download, can use the request_id from nc_requests().

Otherwise, if the user doesn't have "full" access, they must supply an approved request_id to the nc_data_dl() function (e.g., nc_data_dl(request_id = 152000, username = "USER")). Use nc_requests() to see request_ids, filters, and approval status.

Requests for "full" access to additional collections can be made online through the Web Request Form by checking the "Full access?" box in Step 2 of the form.

See also

Examples

# Count all publicly available records:
# \donttest{
nc_count()

#> Without a username, using 'show = "all"'

#> # A tibble: 172 x 4
#>    collection akn_level access     nrecords
#>    <chr>          <int> <chr>         <int>
#>  1 ABATLAS1           5 full         123364
#>  2 ABATLAS2           5 full         201382
#>  3 ABBIRDRECS         5 full         357264
#>  4 ABOWLS             4 by request        0
#>  5 ATBBS              4 by request        0
#>  6 ATOWLS             4 by request    25250
#>  7 BBS                5 full        5735895
#>  8 BBS50              5 full              0
#>  9 BBS50-CAN          5 full        4168740
#> 10 BBS50-MEX          2 no access         0
#> # … with 162 more rows

# }

# Count publicly available records for Manitoba, Canada
# \donttest{
nc_count(region = list(statprov = "MB"))

#> Without a username, using 'show = "all"'

#> Using filters: statprov (MB)

#> # A tibble: 23 x 4
#>    collection    akn_level access     nrecords
#>    <chr>             <int> <chr>         <int>
#>  1 ABBIRDRECS            5 full            357
#>  2 BBS                   5 full          86997
#>  3 BBS50-CAN             5 full         453998
#>  4 CBC                   2 no access     20507
#>  5 CMMN-DET-DMBO         2 no access     88280
#>  6 EBIRD-CA-PR           4 by request  2452054
#>  7 EBIRD-CA-SENS         3 by request     3961
#>  8 EBUTTERFLY            5 full            224
#>  9 GBBC                  4 by request    24393
#> 10 HawkCount             4 by request     1446
#> # … with 13 more rows

# }

# Count all records for all collections user "sample" has access to
if (FALSE) {
nc_count(username = "sample")
}

# Count records with barred owls in Ontario
search_species("barred owl")

#> # A tibble: 15 x 5
#>    species_id scientific_name      english_name     french_name      taxon_group
#>         <int> <chr>                <chr>            <chr>            <chr>      
#>  1       7590 Strix varia          Barred Owl       Chouette rayée   BIRDS      
#>  2       7595 Strix occidentalis … Spotted x Barre… Hybride Chouett… BIRDS      
#>  3      32704 Strix hylophila      Rusty-barred Owl Chouette dryade  BIRDS      
#>  4      32746 Glaucidium cuculoid… Asian Barred Ow… Chevêchette cuc… BIRDS      
#>  5      32750 Glaucidium capense   African Barred … Chevêchette du … BIRDS      
#>  6      32805 Aegotheles bennettii Barred Owlet-Ni… Égothèle de Ben… BIRDS      
#>  7      43347 Glaucidium capense … African Barred … Chevêchette du … BIRDS      
#>  8      43348 Glaucidium capense … African Barred … Chevêchette du … BIRDS      
#>  9      43362 Strix varia [varia … Barred Owl (Nor… Chouette rayée … BIRDS      
#> 10      43363 Strix varia sartorii Barred Owl (Cin… Chouette rayée … BIRDS      
#> 11      43377 Aegotheles bennetti… Barred Owlet-ni… Égothèle de Ben… BIRDS      
#> 12      43378 Aegotheles bennetti… Barred Owlet-ni… Égothèle de Ben… BIRDS      
#> 13      43379 Aegotheles bennetti… Barred Owlet-ni… Égothèle de Ben… BIRDS      
#> 14      45499 Aegotheles bennettii Barred Owlet-ni… Égothèle de Ben… BIRDS      
#> 15      47281 Strix occidentalis/… Spotted/Barred … Chouette tachet… BIRDS      

nc_count(species = 7590, region = list(statprov = "ON"), username = "sample")

#> Using filters: species (7590); statprov (ON)

#> # A tibble: 1 x 4
#>   collection akn_level access nrecords
#>   <chr>          <int> <chr>     <int>
#> 1 SAMPLE2            0 full          1

# Count all records available in the Christmas Bird Count and Breeding Bird
# Survey collections (regardless of user permissions)
nc_count(collections = c("CBC", "BBS"), show = "all", username = "sample")

#> Using filters: collections (CBC, BBS)

#> # A tibble: 2 x 4
#>   collection akn_level access     nrecords
#>   <chr>          <int> <chr>         <int>
#> 1 BBS                5 by request  5735895
#> 2 CBC                2 no access   7570427