Download information about NatureCounts collections

Download the number of records available for different collections filtered by location (if provided). If authorization is provided, the collections are filtered to only those available to the user. Otherwise all collections are returned.

nc_count(collections = NULL, project_ids = NULL, species = NULL,
  years = NULL, doy = NULL, region = NULL, site_type = NULL,
  show = "available", username = NULL, timeout = 30,
  verbose = TRUE)

Arguments

collections
Character vector. The collection codes from which to download data. NULL (default) downloads data from all available collections

project_ids
Character/Numeric vector. The project ids from which to download data. First the collections associated with a project_id are determined, and then data is downloaded for each collection. If both collections and project_ids are supplied, they are combined.

species
Numeric vector. Numeric species ids (see details)

years
Numeric vector. The start/end years of data to download. Can use NA for either start or end, or a single value to return data from a single year.

doy
Character/Numeric vector. The start/end day-of-year to download (1-366 or dates that can be converted to day of year). Can use NA for either start or end

region
List. Named list with one of the following options: country, statprov, subnational2, iba, bcr, utm_squares, bbox. See details

site_type
Character vector. The type of site to return (e.g., IBA).

show
Character. Either "all" or "available". "all" returns counts from all data sources. "available" only returns counts for data available for the username provided. If no username is provided, defaults to "all".

username
Character vector. Username for http://naturecounts.ca. If provided, the user will be prompted for a password. If left NULL, only public collections will be returned.

timeout
Numeric. Number of seconds before connecting to the server times out.

verbose
Logical. Show messages?

Value

Data frame

NatureCounts account

All public data is available without a username/password. However, to access private/semi-public projects/collections you must sign up for a free NatureCounts account and register for the projects you'd like to access. See the request_id section for more information.

Species ids (species)

Numeric species id codes can determined from the functions search_species() or search_species_code(). See also the species codes for more information.

Day of Year (doy)

The format for day of year (doy) is fairly flexible and can be a whole number between 1 and 366 or anything recognized by lubridate-package's ymd() function. However, it must have the order of year, month, day. Note that year is ignored when converting to day of year, except that it will result in a 1 day offset for leap years.

Regions (region)

Regions are defined by codes reflecting the country, state/province, subnational (level 2), Important Bird Areas (IBA), and Bird Conservation Regions (BCR) (see search_region() for codes). They can also be defined by providing specific UTM squares to download or a bounding box area which specifies the min/max longitude and min/max latitude (bbox). See the article on regional filters for more information.

Examples

# Count all publicly available records:

nc_count()

#> Without a username, using 'show = "all"'

#>                collection nrecords
#> 1                ABATLAS1   123364
#> 2                ABATLAS2   201398
#> 3              ABBIRDRECS   357264
#> 4                  ATOWLS      367
#> 5                     BBS  5735895
#> 6               BBS50-CAN  3219534
#> 7           BBS50-US-EAST  8067776
#> 8           BBS50-US-WEST  8701116
#> 9           BCATLAS1BE_DO    41412
#> 10         BCATLAS1BE_RAW   344433
#> 11        BCATLAS1BE_SUMM   165485
#> 12             BCATLAS1PC   260647
#> 13             BCATLAS1RC    21884
#> 14                BCBEACH     7263
#> 15                  BCCWS   346475
#> 16                   BCMA    11291
#> 17                 BCOWLS    27149
#> 18                  BGRMM      476
#> 19              BIRDSOFBC   498444
#> 20                    CBC  7570427
#> 21         CMMN-BAND-LPBO  1168735
#> 22           CMMN-DET-ABO    85569
#> 23          CMMN-DET-ACBO    28386
#> 24           CMMN-DET-BBO    72759
#> 25          CMMN-DET-BPBO    88397
#> 26          CMMN-DET-DMBO    88280
#> 27           CMMN-DET-HBO   271674
#> 28           CMMN-DET-IBS    29736
#> 29          CMMN-DET-IPBO    49741
#> 30          CMMN-DET-LMBO    57432
#> 31          CMMN-DET-LPBO  1073798
#> 32         CMMN-DET-LSLBO   114040
#> 33          CMMN-DET-MGBO    80063
#> 34          CMMN-DET-MIBO    10804
#> 35           CMMN-DET-MNO    69732
#> 36           CMMN-DET-OOT   106146
#> 37         CMMN-DET-PEPBO   131973
#> 38          CMMN-DET-PIBO   116763
#> 39          CMMN-DET-RPBO   131972
#> 40          CMMN-DET-TCBO   136565
#> 41         CMMN-DET-TLBBS    28037
#> 42          CMMN-DET-TLBO    29722
#> 43        CMMN-DET-TTPBRS    97635
#> 44          CMMN-DET-VLBO    52723
#> 45          CMMN-DET-WPBO    63251
#> 46         CMMN-LPBO-MOBU     2863
#> 47            EBIRD-CA-AT  3901226
#> 48            EBIRD-CA-BC 11218063
#> 49            EBIRD-CA-NO   753993
#> 50            EBIRD-CA-ON 20104013
#> 51            EBIRD-CA-PR  5898226
#> 52            EBIRD-CA-QC 15818301
#> 53          EBIRD-CA-SENS    26678
#> 54             EBUTTERFLY    26192
#> 55                   GBBC   431613
#> 56              HAWKCOUNT   112673
#> 57                   IMMP       43
#> 58                IMMP_A2       24
#> 59                IMMP_BW       19
#> 60                   IMQC      761
#> 61          MBATLAS1BE_DO    93876
#> 62         MBATLAS1BE_RAW   326429
#> 63        MBATLAS1BE_SUMM   154856
#> 64             MBATLAS1PC   306094
#> 65             MBATLAS1RC    15622
#> 66            MBBA1BE_RAW   142920
#> 67           MBBA1BE_SUMM    93053
#> 68            MBBA2BE_RAW   263809
#> 69           MBBA2BE_SUMM   114616
#> 70                MBBA2PC   127447
#> 71            MBBA2PC_SAR       13
#> 72                MBBA2RC    14213
#> 73                 MBOWLS    23561
#> 74                   MEXU     3251
#> 75                   MLMP      802
#> 76                     MM     4343
#> 77               MMPBIRDS   425826
#> 78               MMPFROGS    88438
#> 79           MONARCHWATCH   145665
#> 80            Naturalista       30
#> 81              NESTWATCH   413415
#> 82            OBBA1BE_RAW   407287
#> 83           OBBA1BE_SUMM   205816
#> 84            OBBA2BE_RAW   541743
#> 85           OBBA2BE_SUMM   256702
#> 86                OBBA2PC   592425
#> 87            OBBA2PC_SAR       31
#> 88                OBBA2RC    27970
#> 89                   OBFS   324560
#> 90              OISEAUXQC   666963
#> 91                 ONOWLS    40449
#> 92                    PFW  3841194
#> 93            PFW-US-EAST  8651432
#> 94            PFW-US-WEST  5631851
#> 95                   PMMM        5
#> 96             PRISM-ACSS   117227
#> 97              PRISM-OSS    47542
#> 98              PRISM-PSS     3305
#> 99    QCATLAS_NORTH_BE_DO    19747
#> 100  QCATLAS_NORTH_BE_RAW    35599
#> 101 QCATLAS_NORTH_BE_SUMM    23601
#> 102      QCATLAS_NORTH_PC    21579
#> 103        QCATLAS1BE_RAW   198816
#> 104       QCATLAS1BE_SUMM   136333
#> 105         QCATLAS2BE_DO   283034
#> 106        QCATLAS2BE_RAW   601540
#> 107       QCATLAS2BE_SUMM   231775
#> 108            QCATLAS2PC   341902
#> 109            QCATLAS2RC    39390
#> 110                QCOWLS    11195
#> 111           RCBIOTABASE    12598
#> 112              SAR-RSHA    27696
#> 113         SKATLAS1BE_DO    99872
#> 114       SKATLAS1BE_SUMM    42203
#> 115            SKATLAS1PC    27916
#> 116            SWIFTWATCH    10624
#> 117                  WMMM     8981
#> 118                  WPWI     3012

# Count publicly available records for Manitoba, Canada

nc_count(region = list(statprov = "MB"))

#> Without a username, using 'show = "all"'

#> Using filters: statprov (MB)

#>         collection nrecords
#> 1       ABBIRDRECS      357
#> 2              BBS    86997
#> 3        BBS50-CAN   353755
#> 4              CBC    20507
#> 5    CMMN-DET-DMBO    88280
#> 6      EBIRD-CA-PR  1602592
#> 7    EBIRD-CA-SENS     2471
#> 8       EBUTTERFLY      225
#> 9             GBBC    24393
#> 10       HAWKCOUNT     1446
#> 11            IMQC       26
#> 12   MBATLAS1BE_DO    93876
#> 13  MBATLAS1BE_RAW   326429
#> 14 MBATLAS1BE_SUMM   154856
#> 15      MBATLAS1PC   306094
#> 16      MBATLAS1RC    15622
#> 17          MBOWLS    23561
#> 18              MM       53
#> 19        MMPBIRDS       60
#> 20    MONARCHWATCH       24
#> 21       NESTWATCH     4553
#> 22       OISEAUXQC        2
#> 23             PFW   110414

# Count all records for all collections user "sample" has access to

# NOT RUN {
nc_count(username = "sample")
# }

# Count records with barred owls in Ontario
search_species("barred owl")

#>    species_id                            scientific_name
#> 1        7590                                Strix varia
#> 2        7595                 Strix occidentalis x varia
#> 3       32704                            Strix hylophila
#> 4       32746                      Glaucidium cuculoides
#> 5       32750                         Glaucidium capense
#> 6       32805                       Aegotheles bennettii
#> 7       43347              Glaucidium capense scheffleri
#> 8       43348       Glaucidium capense capense/ngamiense
#> 9       43362                  Strix varia [varia Group]
#> 10      43363                       Strix varia sartorii
#> 11      43377 Aegotheles bennettii bennettii/wiedenfeldi
#> 12      43378             Aegotheles bennettii terborghi
#> 13      43379              Aegotheles bennettii plumifer
#> 14      45499                       Aegotheles bennettii
#>                           english_name
#> 1                           Barred Owl
#> 2        Spotted x Barred Owl (hybrid)
#> 3                     Rusty-barred Owl
#> 4                   Asian Barred Owlet
#> 5                 African Barred Owlet
#> 6                Barred Owlet-Nightjar
#> 7  African Barred Owlet (Spot-fronted)
#> 8   African Barred Owlet (Bar-fronted)
#> 9                Barred Owl (Northern)
#> 10              Barred Owl (Cinereous)
#> 11      Barred Owlet-nightjar (Barred)
#> 12   Barred Owlet-nightjar (Diamond's)
#> 13       Barred Owlet-nightjar (Dwarf)
#> 14               Barred Owlet-nightjar
#>                                       french_name taxon_group
#> 1                                  Chouette rayée       BIRDS
#> 2            Hybride Chouette tachetée x C. rayée       BIRDS
#> 3                                 Chouette dryade       BIRDS
#> 4                           Chevêchette cuculoïde       BIRDS
#> 5                              Chevêchette du Cap       BIRDS
#> 6                             Égothèle de Bennett       BIRDS
#> 7                 Chevêchette du Cap (scheffleri)       BIRDS
#> 8       Chevêchette du Cap [capense ou ngamiense]       BIRDS
#> 9                          Chouette rayée [varia]       BIRDS
#> 10                      Chouette rayée (sartorii)       BIRDS
#> 11 Égothèle de Bennett [bennettii ou wiedenfeldi]       BIRDS
#> 12                Égothèle de Bennett (terborghi)       BIRDS
#> 13                 Égothèle de Bennett (plumifer)       BIRDS
#> 14                            Égothèle de Bennett       BIRDS

nc_count(species = 7590, region = list(statprov = "ON"), username = "sample")

#> Using filters: species (7590); statprov (ON)

#>   access  collection nrecords
#> 1    yes RCBIOTABASE       44

# Count all records available in the Christmas Bird Count and Breeding Bird
# Survey collections (regardless of user permissions)

nc_count(collections = c("CBC", "BBS"), show = "all", username = "sample")

#> Using filters: collections (CBC, BBS)

#>   access collection nrecords
#> 1     no        BBS  5735895
#> 2     no        CBC  7570427