Download information about NatureCounts collections

Download the number of records available for different collections filtered by location (if provided). If authorization is provided, the collections are filtered to only those available to the user. Otherwise all collections are returned.

nc_count(collections = NULL, species = NULL, years = NULL,
  doy = NULL, region = NULL, site_type = NULL, show = "available",
  username = NULL, verbose = TRUE)

Arguments

collections
Character vector. The collection codes from which to download data. NULL (default) downloads data from all available collections

species
Numeric vector. Numeric species ids (see details)

years
Numeric vector. The start/end years of data to download. Can use NA for either start or end, or a single value to return data from a single year.

doy
Character/Numeric vector. The start/end day-of-year to download (1-366 or dates that can be converted to day of year). Can use NA for either start or end

region
List. Named list with one of the following options: country, statprov, subnational2, iba, bcr, utm_squares, bbox. See details

site_type
Character vector. The type of site to return (e.g., IBA).

show
Character. Either "all" or "available". "all" returns counts from all data sources. "available" only returns counts for data available (public or accessible with the token provided).

username
Character vector. Username for http://naturecounts.ca. If provided, the user will be prompted for a password. If left NULL, only public collections will be returned.

Value

Data frame

NatureCounts account

All public data is available without a username/password. However, to access private/semi-public projects/collections you must sign up for a free NatureCounts account and register for the projects you'd like to access.

Species ids (species)

Numeric species id codes can determined from the functions species_search() or species_code_search(). See also the species codes (https://birdstudiescanada.github.io/naturecounts/articles/species-codes.html) for more information.

Day of Year (doy)

The format for day of year (doy) is fairly flexible and can be a whole number between 1 and 366 or anything recognized by lubridate-package's ymd() function. However, it must have the order of year, month, day. Note that year is ignored when converting to day of year, except that it will result in a 1 day offset for leap years.

Regions (region)

Regions are defined by codes reflecting the country, state/province, subnational (level 2), Important Bird Areas (IBA), and Bird Conservation Regions (BCR) (see region_search() for codes). They can also be defined by providing specific UTM squares to download or a bounding box area which specifyings the min/max longitude and min/max latitude (bbox). See the article on regional filters (http://BirdStudiesCanada.github.io/naturecounts/articles/region-codes.html) for more information.

Examples

# Count all publicly available records:

nc_count()

#>     collection nrecords
#> 1     ABATLAS1   123364
#> 2     ABATLAS2   201398
#> 3   ABBIRDRECS   357264
#> 4        BGRMM      476
#> 5      IMMP_A2       24
#> 6      IMMP_BW       19
#> 7         IMQC      761
#> 8         MLMP      425
#> 9           MM     4343
#> 10        PMMM        5
#> 11 RCBIOTABASE    12598

# Count publicly available records for Manitoba, Canada

nc_count(region = list(statprov = "MB"))

#> Using filters: statprov (MB)

#>   collection nrecords
#> 1 ABBIRDRECS      357
#> 2       IMQC       26
#> 3         MM       53

# Count all records for all collections you have access to

# NOT RUN {
nc_count(username = "YOUR_USERNAME")
# }

# Count all public records with barred owls in Ontario
species_search("barred owl")

#>    species_id                            scientific_name
#> 1        7590                                Strix varia
#> 2        7595                 Strix occidentalis x varia
#> 3       32704                            Strix hylophila
#> 4       32746                      Glaucidium cuculoides
#> 5       32750                         Glaucidium capense
#> 6       32805                       Aegotheles bennettii
#> 7       43347              Glaucidium capense scheffleri
#> 8       43348       Glaucidium capense capense/ngamiense
#> 9       43362                  Strix varia [varia Group]
#> 10      43363                       Strix varia sartorii
#> 11      43377 Aegotheles bennettii bennettii/wiedenfeldi
#> 12      43378             Aegotheles bennettii terborghi
#> 13      43379              Aegotheles bennettii plumifer
#> 14      45499                       Aegotheles bennettii
#>                           english_name
#> 1                           Barred Owl
#> 2        Spotted x Barred Owl (hybrid)
#> 3                     Rusty-barred Owl
#> 4                   Asian Barred Owlet
#> 5                 African Barred Owlet
#> 6                Barred Owlet-Nightjar
#> 7  African Barred Owlet (Spot-fronted)
#> 8   African Barred Owlet (Bar-fronted)
#> 9                Barred Owl (Northern)
#> 10              Barred Owl (Cinereous)
#> 11      Barred Owlet-nightjar (Barred)
#> 12   Barred Owlet-nightjar (Diamond's)
#> 13       Barred Owlet-nightjar (Dwarf)
#> 14               Barred Owlet-nightjar
#>                                       french_name taxon_group
#> 1                                  Chouette rayée       BIRDS
#> 2            Hybride Chouette tachetée x C. rayée       BIRDS
#> 3                                 Chouette dryade       BIRDS
#> 4                           Chevêchette cuculoïde       BIRDS
#> 5                              Chevêchette du Cap       BIRDS
#> 6                             Égothèle de Bennett       BIRDS
#> 7                 Chevêchette du Cap (scheffleri)       BIRDS
#> 8       Chevêchette du Cap [capense ou ngamiense]       BIRDS
#> 9                          Chouette rayée [varia]       BIRDS
#> 10                      Chouette rayée (sartorii)       BIRDS
#> 11 Égothèle de Bennett [bennettii ou wiedenfeldi]       BIRDS
#> 12                Égothèle de Bennett (terborghi)       BIRDS
#> 13                 Égothèle de Bennett (plumifer)       BIRDS
#> 14                            Égothèle de Bennett       BIRDS

nc_count(species = 7590, region = list(statprov = "ON"))

#> Using filters: species (7590); statprov (ON)

#>    collection nrecords
#> 1 RCBIOTABASE       44

# Count records available in the Christmas Bird Count and Breeding Bird
# Survey collections (regardless of permissions)

nc_count(collections = c("CBC", "BBS"), show = "all")

#> Using filters: collections (CBC, BBS)

#>   collection nrecords
#> 1        BBS  5735895
#> 2        CBC  7570427