Download NatureCounts data records

Download data records from various collections filtered by various options. In order to ease the load on the server, note that only three of collections, species, years, doy, region, and site_type can be used in any one request. See the vignette for filtering your data after download for more options: vignette("filtering_data", package = "naturecounts").

nc_data_dl(collections = NULL, species = NULL, years = NULL,
  doy = NULL, region = NULL, site_type = NULL,
  fields_set = "minimum", fields = NULL, token = NULL,
  sql_db = NULL, verbose = TRUE)

Arguments

collections
Character vector. The collection codes from which to download data. NULL (default) downloads data from all available collections

species
Numeric vector. Numeric species ids (see details)

years
Numeric vector. The start/end years of data to download. Can use NA for either start or end, or a single value to return data from a single year.

doy
Character/Numeric vector. The start/end day-of-year to download (1-366 or dates that can be converted to day of year). Can use NA for either start or end

region
List. Named list with one of the following options: country, statprov, subnational2, iba, bcr, utm_squares, bbox. See details

site_type
Character vector. The type of site to return (e.g., IBA).

fields_set
Charcter. Set of fields/columns to download. See details.

fields
Character vector. If fields_set = custom, which fields/columns to download. See details

token
Character vector. Authorization token, otherwise only public data is accessible

sql_db
Character vector. Name and location of SQLite database to either create or add to

verbose
Logical. Display progress messages?

Value

Data frame

Data Fields/Columns

By default data is downloaded with the minimum set of fields/columns. However, for more advanced applications, users may wish to specify which fields/columns to return. The Bird Monitoring Data Exchange (BMDE) schema keeps track of variables used to augment observation data. There are different versions reflecting different collections of variables which can be specified for download in one of four ways:

fields_set can be a specific shorthand reflecting a BMDE version: core, extended or minimum (default). See meta_bmde_versions() to see which BMDE version the shorthand refers to.

fields_set can be default which uses the default BMDE version for a particular collection (note that if you download more than one collection, the field sets will expand to cover all fields/columns in the combined collections)

fields_set can be the exact BMDE version. See meta_bmde_versions() for options.

fields_set can be custom and the fields argument can be a character vector specifying the exact fields/columns to return. See meta_bmde_fields()) for potential fields values.

Note that in all cases there are a set of fields/columns that are always returned, no matter what fields_set is used.

Species ids (species)

Numeric species id codes can determined from the functions species_search() or species_code_search(). See also the vignette for more information vignette("species-codes", package = "naturecounts").

Day of Year (doy)

The format for day of year (doy) is fairly flexible and can be a whole number between 1 and 366 or anything recognized by lubridate-package's ymd() function. However, it must have the order of year, month, day. Note that year is ignored when converting to day of year, except that it will result in a 1 day offset for leap years.

Regions (region)

Regions are defined by codes reflecting the country, state/province, subnational (level 2), Important Bird Areas (IBA), and Bird Conservation Regions (BCR) (see region_search() for codes). They can also be defined by providing specific UTM squares to download or a bounding box area which specifyings the min/max longitude and min/max latitude (bbox). See the article on regional filtering for more detail: http://BirdStudiesCanada.github.io/naturecounts/articles/region.html

Examples

# All observations part of the RCBIOTABASE collection
rcbio <- nc_data_dl(collection = "RCBIOTABASE")

#> Collecting available records...

#>    collection nrecords
#> 1 RCBIOTABASE    12598

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 5000 / 12598

# Observations of black-capped chickadees from RCBIOTABASE collection in 2010
species_search("black-capped chickadee") # Find the species_id

#>   species_id                           scientific_name
#> 1      14280                      Poecile atricapillus
#> 2      40668       Poecile carolinensis x atricapillus
#> 3      40669         Poecile carolinensis/atricapillus
#> 4      44466 Poecile atricapillus x Baeolophus bicolor
#>                                        english_name
#> 1                            Black-capped Chickadee
#> 2        Carolina x Black-capped Chickadee (hybrid)
#> 3                   Carolina/Black-capped Chickadee
#> 4 Black-capped Chickadee x Tufted Titmouse (hybrid)
#>                                     french_name taxon_group
#> 1                          Mésange à tête noire       BIRDS
#> 2 Hybride Mésange de Caroline x M. à tête noire       BIRDS
#> 3        Mésange de Caroline ou M. à tête noire       BIRDS
#> 4    Hybride Mésange à tête noire x M. bicolore       BIRDS

bcch <- nc_data_dl(collection = "RCBIOTABASE", species = 14280,
                   years = 2010)

#> Collecting available records...

#>    collection nrecords
#> 1 RCBIOTABASE        7

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 7 / 7

# All public bcch observations since 2015
bcch <- nc_data_dl(species = 14280, years = c(2015, NA))

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>    collection nrecords
#> 1 RCBIOTABASE        2

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 2 / 2

bcch <- nc_data_dl(species = 14280, doy = c(200, 300))

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>    collection nrecords
#> 1    ABATLAS1       55
#> 2    ABATLAS2      565
#> 3  ABBIRDRECS      223
#> 4 RCBIOTABASE       17

#> 
#> Downloading records for each collection:

#>   ABATLAS1

#>     Records 1 to 55 / 55

#>   ABATLAS2

#>     Records 1 to 565 / 565

#>   ABBIRDRECS

#>     Records 1 to 223 / 223

#>   RCBIOTABASE

#>     Records 1 to 17 / 17

bcch <- nc_data_dl(species = 14280,
                   region = list(bbox = c(left = -145, bottom = 45,
                                          right = -100, top = 60)))

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>   collection nrecords
#> 1   ABATLAS1     1314
#> 2   ABATLAS2     3773
#> 3 ABBIRDRECS     2931

#> 
#> Downloading records for each collection:

#>   ABATLAS1

#>     Records 1 to 1314 / 1314

#>   ABATLAS2

#>     Records 1 to 3773 / 3773

#>   ABBIRDRECS

#>     Records 1 to 2931 / 2931

# All moose observations with public access
species_search("moose")

#>   species_id scientific_name english_name french_name taxon_group
#> 1     133990     Alces alces        Moose        <NA>     MAMMALS

moose <- nc_data_dl(species = 133990)

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>    collection nrecords
#> 1 RCBIOTABASE        2

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 2 / 2

# Different fields/columns
moose <- nc_data_dl(species = 133990, fields_set = "core")

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>    collection nrecords
#> 1 RCBIOTABASE        2

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 2 / 2

moose <- nc_data_dl(species = 133990, fields_set = "custom",
                    fields = c("Locality", "AllSpeciesReported"))

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>    collection nrecords
#> 1 RCBIOTABASE        2

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 2 / 2