Download NatureCounts data records

Download data records from various collections filtered by date and location (if provided). An authorization token is required.

nc_data_dl(collections = NULL, species = NULL, start_date = NULL,
  end_date = NULL, location = NULL, country = NULL,
  statprov = NULL, token = NULL, sql_db = NULL, verbose = TRUE)

Arguments

collections
Character vector. The collection codes from which to download data. NULL (default) downloads data from all available collections

species
Numeric vector. Numeric species ids (see details)

start_date
Character. The starting date of data to download. See details for format

end_date
Character. The end date of data to download. See details for format

location
NOT IMPLEMENTED

country
Character vector. Filter data to specific countries codes

statprov
Character vector. Filter data to specific states/province codes

token
Character vector. Authorization token, otherwise only public data is accessible.

sql_db
Character vector. Name and location of SQLite database to either create or add to.

verbose
Logical. Display progress messages?

Value

Data frame

Details

Numeric species id codes can determined from the functions species_search() or species_code_search().

The format of start/end dates is fairly flexible and can be anything recognized by lubridate-package's ymd() function. However, it must have the order of year, month, day. Month and day are option: It can be year and month (e.g., "2000 May"), or simply year (e.g., 2000).

Examples

# All observations part of the RCBIOTABASE collection
rcbio <- nc_data_dl(collection = "RCBIOTABASE")

#> Collecting available records...

#>    collection nrecords
#> 1 RCBIOTABASE    12598

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 5000 / 12598

#>     Records 5001 to 10000 / 12598

#>     Records 10001 to 12598 / 12598

# Observations of black-capped chickadees from RCBIOTABASE collection in 2010
species_search("black-capped chickadee") # Find the species_id

#>   species_id                           scientific_name
#> 1      14280                      Poecile atricapillus
#> 2      40668       Poecile carolinensis x atricapillus
#> 3      40669         Poecile carolinensis/atricapillus
#> 4      44466 Poecile atricapillus x Baeolophus bicolor
#>                                        english_name
#> 1                            Black-capped Chickadee
#> 2        Carolina x Black-capped Chickadee (hybrid)
#> 3                   Carolina/Black-capped Chickadee
#> 4 Black-capped Chickadee x Tufted Titmouse (hybrid)
#>                                     french_name taxon_group
#> 1                          Mésange à tête noire       BIRDS
#> 2 Hybride Mésange de Caroline x M. à tête noire       BIRDS
#> 3        Mésange de Caroline ou M. à tête noire       BIRDS
#> 4    Hybride Mésange à tête noire x M. bicolore       BIRDS

bcch <- nc_data_dl(collection = "RCBIOTABASE", species = 14280,
                   start_date = 2010, end_date = 2010)

#> Collecting available records...

#>    collection nrecords
#> 1 RCBIOTABASE        7

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 7 / 7

# All moose observations with public access
species_search("moose")

#>   species_id scientific_name english_name french_name taxon_group
#> 1     133990     Alces alces        Moose        <NA>     MAMMALS

moose <- nc_data_dl(species = 133990)

#> Collecting available records...

#> Not all collections have data that match these filters ()

#>    collection nrecords
#> 1 RCBIOTABASE        2

#> 
#> Downloading records for each collection:

#>   RCBIOTABASE

#>     Records 1 to 2 / 2