Download list of accessible collections

Returns a list of collections accessible by 'username'.

nc_permissions(username = NULL, timeout = 60)

Arguments

username

Character vector. Username for http://naturecounts.ca. If provided, the user will be prompted for a password. If left NULL, only public collections will be returned.

timeout

Numeric. Number of seconds before connecting to the server times out.

NatureCounts account

All public data is available with a username/password (sign up for a free NatureCounts account). However, to access private/semi-public projects/collections you must request access. See the Access and request_ids section for more information.

Examples

nc_permissions()
#>  [1] "ABATLAS1"          "ABATLAS2"          "ABBIRDRECS"       
#>  [4] "BBS"               "BBS50"             "BBS50-CAN"        
#>  [7] "BBS50-US-EAST"     "BBS50-US-WEST"     "BCATLAS1BE_DO"    
#> [10] "BCATLAS1BE_RAW"    "BCATLAS1BE_SUMM"   "BCATLAS1PC"       
#> [13] "BCMA"              "BGRMM"             "CMMN-DET-HBO"     
#> [16] "CMMN-DET-RUTH"     "CMMN-DET-SELK"     "CMMN-DET-VLBO"    
#> [19] "EBUTTERFLY"        "IMMP"              "IMMP_A2"          
#> [22] "IMMP_BW"           "IMQC"              "MBATLAS1BE_DO"    
#> [25] "MBATLAS1BE_RAW"    "MBATLAS1BE_SUMM"   "MBATLAS1PC"       
#> [28] "MBBA1BE_RAW"       "MBBA1BE_SUMM"      "MBBA2BE_RAW"      
#> [31] "MBBA2BE_SUMM"      "MBBA2PC"           "MEXU"             
#> [34] "MLMP"              "MM"                "MONARCHWATCH"     
#> [37] "NATURALISTA"       "NESTWATCH_KMARTIN" "NRCS_KMARTIN"     
#> [40] "OBBA1BE_RAW"       "OBBA1BE_SUMM"      "OBBA2BE_RAW"      
#> [43] "OBBA2BE_SUMM"      "OBBA2PC"           "OBFS"             
#> [46] "PFW"               "PMMM"              "QCATLAS1BE_RAW"   
#> [49] "QCATLAS1BE_SUMM"   "QCATLAS2BE_DO"     "QCATLAS2BE_RAW"   
#> [52] "QCATLAS2BE_SUMM"   "QCATLAS2PC"        "RCBIOTABASE"      
#> [55] "WMMM"              "WPWI"             
nc_permissions(username = "sample")
#> [1] "SAMPLE1" "SAMPLE2"