Find country, state/province, subnational2, IBA, or BCR codes

Search for the correct codes to identify countries, states/provinces, subnational2 areas, Important Bird Areas (IBA), or Bird Conservation Regions (BCR). These are then used in the nc_data_dl() and nc_count() functions.

search_region(name = NULL, type = "country")

Arguments

name

Character. The location name to search for

type

Character. One of "country", "statprov", "subnational2", "iba", or "bcr". The type of information to return.

Value

A data frame with the relevant codes and other information

Details

region_search() is deprecated in favour of search_region()

Examples

search_region("Mexico", type = "country")   # MX
#> # A tibble: 1 × 3
#>   country_code country_name country_name_fr
#>   <chr>        <chr>        <chr>          
#> 1 MX           Mexico       Mexique        

search_region("Yucatan", type = "statprov") # Yucatán
#> # A tibble: 1 × 5
#>   country_code statprov_code statprov_name_es statprov_name statprov_name_fr
#>   <chr>        <chr>         <chr>            <chr>         <chr>           
#> 1 MX           YP            Yucatán          Yucatán       Yucatán         
search_region("Alberta", type = "statprov") # AB
#> # A tibble: 1 × 5
#>   country_code statprov_code statprov_name_es statprov_name statprov_name_fr
#>   <chr>        <chr>         <chr>            <chr>         <chr>           
#> 1 CA           AB            Alberta          Alberta       Alberta         

search_region("Edmonton", type = "subnational2") # CA.AB.11
#> # A tibble: 1 × 5
#>   country_code statprov_code subnational2_code subnational2_name         ebird…¹
#>   <chr>        <chr>         <chr>             <chr>                     <chr>  
#> 1 CA           AB            CA.AB.11          Division No. 11 - Edmont… CA-AB-…
#> # … with abbreviated variable name ¹​ebird_code
search_region("Brandon", type = "subnational2")  # CA.MB.07
#> # A tibble: 1 × 5
#>   country_code statprov_code subnational2_code subnational2_name         ebird…¹
#>   <chr>        <chr>         <chr>             <chr>                     <chr>  
#> 1 CA           MB            CA.MB.07          Division No. 7 - Brandon… CA-MB-…
#> # … with abbreviated variable name ¹​ebird_code

search_region("hays reservoir", type = "iba") # AB075
#> # A tibble: 1 × 13
#>   iba_site iba_n…¹ latit…² statp…³ area_ha alt_min iba_n…⁴ neare…⁵ bcr   ncc_r…⁶
#>   <chr>    <chr>     <dbl> <chr>     <dbl>   <dbl> <chr>   <chr>   <chr> <chr>  
#> 1 AB075    Hays R…    50.1 AB         8.99     785 Hays R… Hays    11    26     
#> # … with 3 more variables: alt_max <dbl>, longitude <dbl>, status <chr>, and
#> #   abbreviated variable names ¹​iba_name_fr, ²​latitude, ³​statprov, ⁴​iba_name,
#> #   ⁵​nearest_town, ⁶​ncc_region
search_region("rainforest", type = "bcr")     # 5
#> # A tibble: 1 × 4
#>     bcr bcr_name                    bcr_name_es                        bcr_nam…¹
#>   <int> <chr>                       <chr>                              <chr>    
#> 1     5 Northern Pacific Rainforest Bosque Lluvioso del Pacífico Norte Forêts p…
#> # … with abbreviated variable name ¹​bcr_name_fr

# Show all codes
search_region(type = "country")
#> # A tibble: 39 × 3
#>    country_code country_name country_name_fr
#>    <chr>        <chr>        <chr>          
#>  1 AQ           Antarctica   Antarctique    
#>  2 AR           Argentina    Argentine      
#>  3 BB           Barbados     Barbade        
#>  4 BM           Bermuda      Bermudes       
#>  5 BR           Brazil       Brésil         
#>  6 BS           Bahamas      Bahamas        
#>  7 BZ           Belize       Belize         
#>  8 CA           Canada       Canada         
#>  9 CL           Chile        Chili          
#> 10 CO           Colombia     Colombie       
#> # … with 29 more rows
search_region(type = "statprov")
#> # A tibble: 96 × 5
#>    country_code statprov_code statprov_name_es          statprov_name    statp…¹
#>    <chr>        <chr>         <chr>                     <chr>            <chr>  
#>  1 CA           AB            Alberta                   Alberta          Alberta
#>  2 CA           BC            Columbia Británica        British Columbia Colomb…
#>  3 CA           MB            Manitoba                  Manitoba         Manito…
#>  4 CA           NB            Nuevo Brunswick           New Brunswick    Nouvea…
#>  5 CA           NL            Terranova y Labrador      Newfoundland an… Terre-…
#>  6 CA           NS            Nueva Escocia             Nova Scotia      Nouvel…
#>  7 CA           NT            Territorios del Noroeste  Northwest Terri… Territ…
#>  8 CA           NU            Nunavut                   Nunavut          Nunavut
#>  9 CA           ON            Ontario                   Ontario          Ontario
#> 10 CA           PE            Isla del Príncipe Eduardo Prince Edward I… Île-du…
#> # … with 86 more rows, and abbreviated variable name ¹​statprov_name_fr
search_region(type = "subnational2")
#> # A tibble: 5,310 × 5
#>    country_code statprov_code subnational2_code subnational2_name        ebird…¹
#>    <chr>        <chr>         <chr>             <chr>                    <chr>  
#>  1 CA           AB            CA.AB.01          Division No. 1 - Medici… CA-AB-…
#>  2 CA           AB            CA.AB.02          Division No. 2 - Lethbr… CA-AB-…
#>  3 CA           AB            CA.AB.03          Division No. 3 - Fort M… CA-AB-…
#>  4 CA           AB            CA.AB.04          Division No. 4 - Hanna   CA-AB-…
#>  5 CA           AB            CA.AB.05          Division No. 5 - Drumhe… CA-AB-…
#>  6 CA           AB            CA.AB.06          Division No. 6 - Calgary CA-AB-…
#>  7 CA           AB            CA.AB.07          Division No. 7 - Stettl… CA-AB-…
#>  8 CA           AB            CA.AB.08          Division No. 8 - Red De… CA-AB-…
#>  9 CA           AB            CA.AB.09          Division No. 9 - Rocky … CA-AB-…
#> 10 CA           AB            CA.AB.10          Division No. 10 - Camro… CA-AB-…
#> # … with 5,300 more rows, and abbreviated variable name ¹​ebird_code
search_region(type = "iba")
#> # A tibble: 599 × 13
#>    iba_s…¹ iba_n…² latit…³ statp…⁴ area_ha alt_min iba_n…⁵ neare…⁶ bcr   ncc_r…⁷
#>    <chr>   <chr>     <dbl> <chr>     <dbl>   <dbl> <chr>   <chr>   <chr> <chr>  
#>  1 AB001   Beaver…    53.5 AB        208.      668 Beaver… Tofield 11    66     
#>  2 AB002   Peace-…    58.7 AB       7585.      250 Peace-… Fort C… 06    00     
#>  3 AB003   Lesser…    55.4 AB       2019.      570 Lesser… Slave … 06    00     
#>  4 AB004   Milk R…    49.1 AB        335.      834 Milk R… Milk R… 11    26     
#>  5 AB006   Lakela…    54.7 AB        741.      500 Lakela… Lac la… 06    00     
#>  6 AB007   Suffie…    50.5 AB        461.      610 Suffie… Suffie… 11    26     
#>  7 AB011   Réserv…    49.3 AB         62.9    1100 St. Ma… Cardst… 10    67     
#>  8 AB015   Lake N…    50.4 AB        115.      765 Lake N… Brooks  11    26     
#>  9 AB016   McGreg…    50.3 AB        251.      850 McGreg… Vulcan  11    67     
#> 10 AB022   Little…    51.4 AB         35.7     894 Little… High R… 11    26     
#> # … with 589 more rows, 3 more variables: alt_max <dbl>, longitude <dbl>,
#> #   status <chr>, and abbreviated variable names ¹​iba_site, ²​iba_name_fr,
#> #   ³​latitude, ⁴​statprov, ⁵​iba_name, ⁶​nearest_town, ⁷​ncc_region
search_region(type = "bcr")
#> # A tibble: 67 × 4
#>      bcr bcr_name                       bcr_name_es                      bcr_n…¹
#>    <int> <chr>                          <chr>                            <chr>  
#>  1     1 Aleutian/Bering Sea Islands    Aleutianas/Islas del Mar de Ber… Aleuti…
#>  2     2 Western Alaska                 Alaska Occidental                Wester…
#>  3     3 Arctic Plains And Mountains    Planicies Árticas y Montañas     Plaine…
#>  4     4 Northwestern Interior Forest   Bosque Interior del Noroeste     Forêts…
#>  5     5 Northern Pacific Rainforest    Bosque Lluvioso del Pacífico No… Forêts…
#>  6     6 Boreal Taiga Plains            Planicies de la Taiga Boreal     Taïga …
#>  7     7 Taiga Shield And Hudson Plains Placa de Taiga y Planicies de H… Taïga …
#>  8     8 Boreal Softwood Shield         Placa Boreal de Softwood         Forêts…
#>  9     9 Great Basin                    Gran Cuenca                      Grand …
#> 10    10 Northern Rockies               Rocallosas del Norte             Nord d…
#> # … with 57 more rows, and abbreviated variable name ¹​bcr_name_fr

# \donttest{
# Using the codes
nc_count(region = list(statprov = "AB"), years = 2010)
#> Without a username, using 'show = "all"'
#> Using filters: start_year (2010); end_year (2010); statprov (AB)
#> # A tibble: 13 × 4
#>    collection     akn_level access     nrecords
#>    <chr>              <int> <chr>         <int>
#>  1 BBS                    5 full           9550
#>  2 BBS50-CAN              5 full          31777
#>  3 CBC                    2 no access      1155
#>  4 CMMN-DET-BBO           3 by request     1654
#>  5 CMMN-DET-IBS           3 by request      846
#>  6 CMMN-DET-LSLBO         3 by request     5332
#>  7 EBIRD-CA-PR            4 by request    76793
#>  8 EBIRD-CA-SENS          3 by request      104
#>  9 EBUTTERFLY             5 full              2
#> 10 GBBC                   4 by request     3061
#> 11 NESTWATCH              3 by request        3
#> 12 OISEAUXQC              2 no access        20
#> 13 PFW                    5 full           7033
# }